Crate rustc_llvm

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Reexports

pub use self::IntPredicate::*;

pub use self::RealPredicate::*;

pub use self::TypeKind::*;

pub use self::AtomicRmwBinOp::*;

pub use self::MetadataType::*;

pub use self::CodeGenOptSize::*;

pub use self::CallConv::*;

pub use self::Linkage::*;

Modules

archive_ro 	 [
Experimental
]
A wrapper around LLVM's archive (.a) code

debuginfo 	 [
Experimental
]

diagnostic 	 [
Experimental
]
LLVM diagnostic reports.

Structs

ObjectFile 	 [
Experimental
]

OperandBundleDef 	 [
Experimental
]

SectionIter 	 [
Experimental
]

Enums

ArchiveChild_opaque 	 [
Experimental
]

ArchiveIterator_opaque 	 [
Experimental
]

ArchiveKind 	 [
Experimental
]
LLVMRustArchiveKind

Archive_opaque 	 [
Experimental
]

AsmDialect 	 [
Experimental
]
LLVMRustAsmDialect

AtomicOrdering 	 [
Experimental
]
LLVMAtomicOrdering

AtomicRmwBinOp 	 [
Experimental
]
LLVMAtomicRmwBinOp

Attribute 	 [
Experimental
]
Matches LLVMRustAttribute in rustllvm.h Semantically a subset of the C++ enum llvm::Attribute::AttrKind, though it is not ABI compatible (since it's a C++ enum)

AttributePlace 	 [
Experimental
]

BasicBlock_opaque 	 [
Experimental
]

Builder_opaque 	 [
Experimental
]

CallConv 	 [
Experimental
]
LLVM CallingConv::ID. Should we wrap this?

CodeGenOptLevel 	 [
Experimental
]
LLVMRustCodeGenOptLevel

CodeGenOptSize 	 [
Experimental
]

CodeModel 	 [
Experimental
]
LLVMRustCodeModel

Context_opaque 	 [
Experimental
]

DLLStorageClass 	 [
Experimental
]
LLVMDLLStorageClass

DebugLoc_opaque 	 [
Experimental
]

DiagnosticInfo_opaque 	 [
Experimental
]

DiagnosticKind 	 [
Experimental
]
LLVMRustDiagnosticKind

DiagnosticSeverity 	 [
Experimental
]
LLVMDiagnosticSeverity

ExecutionEngine_opaque 	 [
Experimental
]

FileType 	 [
Experimental
]
LLVMRustFileType

IntPredicate 	 [
Experimental
]
LLVMIntPredicate

LLVMRustResult 	 [
Experimental
]

Linkage 	 [
Experimental
]
LLVMRustLinkage

MemoryBuffer_opaque 	 [
Experimental
]

MetadataType 	 [
Experimental
]
LLVMMetadataType

Metadata_opaque 	 [
Experimental
]

Module_opaque 	 [
Experimental
]

ObjectFile_opaque 	 [
Experimental
]

OperandBundleDef_opaque 	 [
Experimental
]

PassKind 	 [
Experimental
]
LLVMRustPassKind

PassManagerBuilder_opaque 	 [
Experimental
]

PassManager_opaque 	 [
Experimental
]

Pass_opaque 	 [
Experimental
]

RealPredicate 	 [
Experimental
]
LLVMRealPredicate

RelocMode 	 [
Experimental
]
LLVMRelocMode

RustArchiveMember_opaque 	 [
Experimental
]

RustString_opaque 	 [
Experimental
]

SMDiagnostic_opaque 	 [
Experimental
]

SectionIterator_opaque 	 [
Experimental
]

SynchronizationScope 	 [
Experimental
]
LLVMRustSynchronizationScope

TargetData_opaque 	 [
Experimental
]

TargetMachine_opaque 	 [
Experimental
]

Twine_opaque 	 [
Experimental
]

TypeKind 	 [
Experimental
]
LLVMTypeKind

Type_opaque 	 [
Experimental
]

Use_opaque 	 [
Experimental
]

Value_opaque 	 [
Experimental
]

Visibility 	 [
Experimental
]

Constants

False 	 [
Experimental
]

True 	 [
Experimental
]

Functions

AddFunctionAttrStringValue 	 [
Experimental
]

LLVMABIAlignmentOfType⚠ 	 [
Experimental
]
Returns the minimum alignment of a type.

LLVMABISizeOfType⚠ 	 [
Experimental
]
Distance between successive elements in an array of T. Includes ABI padding.

LLVMAddCase⚠ 	 [
Experimental
]

LLVMAddClause⚠ 	 [
Experimental
]

LLVMAddFunction⚠ 	 [
Experimental
]

LLVMAddGlobal⚠ 	 [
Experimental
]

LLVMAddIncoming⚠ 	 [
Experimental
]

LLVMAddNamedMetadataOperand⚠ 	 [
Experimental
]

LLVMAppendBasicBlockInContext⚠ 	 [
Experimental
]

LLVMBasicBlockAsValue⚠ 	 [
Experimental
]

LLVMBuildAShr⚠ 	 [
Experimental
]

LLVMBuildAdd⚠ 	 [
Experimental
]

LLVMBuildAggregateRet⚠ 	 [
Experimental
]

LLVMBuildAlloca⚠ 	 [
Experimental
]

LLVMBuildAnd⚠ 	 [
Experimental
]

LLVMBuildAtomicRMW⚠ 	 [
Experimental
]

LLVMBuildBinOp⚠ 	 [
Experimental
]

LLVMBuildBitCast⚠ 	 [
Experimental
]

LLVMBuildBr⚠ 	 [
Experimental
]

LLVMBuildCast⚠ 	 [
Experimental
]

LLVMBuildCondBr⚠ 	 [
Experimental
]

LLVMBuildExactSDiv⚠ 	 [
Experimental
]

LLVMBuildExtractElement⚠ 	 [
Experimental
]

LLVMBuildExtractValue⚠ 	 [
Experimental
]

LLVMBuildFAdd⚠ 	 [
Experimental
]

LLVMBuildFCmp⚠ 	 [
Experimental
]

LLVMBuildFDiv⚠ 	 [
Experimental
]

LLVMBuildFMul⚠ 	 [
Experimental
]

LLVMBuildFNeg⚠ 	 [
Experimental
]

LLVMBuildFPCast⚠ 	 [
Experimental
]

LLVMBuildFPExt⚠ 	 [
Experimental
]

LLVMBuildFPToSI⚠ 	 [
Experimental
]

LLVMBuildFPToUI⚠ 	 [
Experimental
]

LLVMBuildFPTrunc⚠ 	 [
Experimental
]

LLVMBuildFRem⚠ 	 [
Experimental
]

LLVMBuildFSub⚠ 	 [
Experimental
]

LLVMBuildFree⚠ 	 [
Experimental
]

LLVMBuildGEP⚠ 	 [
Experimental
]

LLVMBuildGlobalString⚠ 	 [
Experimental
]

LLVMBuildGlobalStringPtr⚠ 	 [
Experimental
]

LLVMBuildICmp⚠ 	 [
Experimental
]

LLVMBuildInBoundsGEP⚠ 	 [
Experimental
]

LLVMBuildIndirectBr⚠ 	 [
Experimental
]

LLVMBuildInsertElement⚠ 	 [
Experimental
]

LLVMBuildInsertValue⚠ 	 [
Experimental
]

LLVMBuildIntToPtr⚠ 	 [
Experimental
]

LLVMBuildIsNotNull⚠ 	 [
Experimental
]

LLVMBuildIsNull⚠ 	 [
Experimental
]

LLVMBuildLShr⚠ 	 [
Experimental
]

LLVMBuildLoad⚠ 	 [
Experimental
]

LLVMBuildMul⚠ 	 [
Experimental
]

LLVMBuildNSWAdd⚠ 	 [
Experimental
]

LLVMBuildNSWMul⚠ 	 [
Experimental
]

LLVMBuildNSWNeg⚠ 	 [
Experimental
]

LLVMBuildNSWSub⚠ 	 [
Experimental
]

LLVMBuildNUWAdd⚠ 	 [
Experimental
]

LLVMBuildNUWMul⚠ 	 [
Experimental
]

LLVMBuildNUWNeg⚠ 	 [
Experimental
]

LLVMBuildNUWSub⚠ 	 [
Experimental
]

LLVMBuildNeg⚠ 	 [
Experimental
]

LLVMBuildNot⚠ 	 [
Experimental
]

LLVMBuildOr⚠ 	 [
Experimental
]

LLVMBuildPhi⚠ 	 [
Experimental
]

LLVMBuildPointerCast⚠ 	 [
Experimental
]

LLVMBuildPtrDiff⚠ 	 [
Experimental
]

LLVMBuildPtrToInt⚠ 	 [
Experimental
]

LLVMBuildResume⚠ 	 [
Experimental
]

LLVMBuildRet⚠ 	 [
Experimental
]

LLVMBuildRetVoid⚠ 	 [
Experimental
]

LLVMBuildSDiv⚠ 	 [
Experimental
]

LLVMBuildSExt⚠ 	 [
Experimental
]

LLVMBuildSExtOrBitCast⚠ 	 [
Experimental
]

LLVMBuildSIToFP⚠ 	 [
Experimental
]

LLVMBuildSRem⚠ 	 [
Experimental
]

LLVMBuildSelect⚠ 	 [
Experimental
]

LLVMBuildShl⚠ 	 [
Experimental
]

LLVMBuildShuffleVector⚠ 	 [
Experimental
]

LLVMBuildStore⚠ 	 [
Experimental
]

LLVMBuildStructGEP⚠ 	 [
Experimental
]

LLVMBuildSub⚠ 	 [
Experimental
]

LLVMBuildSwitch⚠ 	 [
Experimental
]

LLVMBuildTrunc⚠ 	 [
Experimental
]

LLVMBuildTruncOrBitCast⚠ 	 [
Experimental
]

LLVMBuildUDiv⚠ 	 [
Experimental
]

LLVMBuildUIToFP⚠ 	 [
Experimental
]

LLVMBuildURem⚠ 	 [
Experimental
]

LLVMBuildUnreachable⚠ 	 [
Experimental
]

LLVMBuildVAArg⚠ 	 [
Experimental
]

LLVMBuildXor⚠ 	 [
Experimental
]

LLVMBuildZExt⚠ 	 [
Experimental
]

LLVMBuildZExtOrBitCast⚠ 	 [
Experimental
]

LLVMCloneModule⚠ 	 [
Experimental
]

LLVMConstAShr⚠ 	 [
Experimental
]

LLVMConstAdd⚠ 	 [
Experimental
]

LLVMConstAnd⚠ 	 [
Experimental
]

LLVMConstArray⚠ 	 [
Experimental
]

LLVMConstBitCast⚠ 	 [
Experimental
]

LLVMConstExtractValue⚠ 	 [
Experimental
]

LLVMConstFAdd⚠ 	 [
Experimental
]

LLVMConstFCmp⚠ 	 [
Experimental
]

LLVMConstFDiv⚠ 	 [
Experimental
]

LLVMConstFMul⚠ 	 [
Experimental
]

LLVMConstFNeg⚠ 	 [
Experimental
]

LLVMConstFPCast⚠ 	 [
Experimental
]

LLVMConstFPToSI⚠ 	 [
Experimental
]

LLVMConstFPToUI⚠ 	 [
Experimental
]

LLVMConstFRem⚠ 	 [
Experimental
]

LLVMConstFSub⚠ 	 [
Experimental
]

LLVMConstICmp⚠ 	 [
Experimental
]

LLVMConstInlineAsm⚠ 	 [
Experimental
]

LLVMConstInt⚠ 	 [
Experimental
]

LLVMConstIntCast⚠ 	 [
Experimental
]

LLVMConstIntGetSExtValue⚠ 	 [
Experimental
]

LLVMConstIntGetZExtValue⚠ 	 [
Experimental
]

LLVMConstIntOfArbitraryPrecision⚠ 	 [
Experimental
]

LLVMConstIntToPtr⚠ 	 [
Experimental
]

LLVMConstLShr⚠ 	 [
Experimental
]

LLVMConstMul⚠ 	 [
Experimental
]

LLVMConstNamedStruct⚠ 	 [
Experimental
]

LLVMConstNeg⚠ 	 [
Experimental
]

LLVMConstNot⚠ 	 [
Experimental
]

LLVMConstNull⚠ 	 [
Experimental
]

LLVMConstOr⚠ 	 [
Experimental
]

LLVMConstPointerCast⚠ 	 [
Experimental
]

LLVMConstPtrToInt⚠ 	 [
Experimental
]

LLVMConstSDiv⚠ 	 [
Experimental
]

LLVMConstSIToFP⚠ 	 [
Experimental
]

LLVMConstSRem⚠ 	 [
Experimental
]

LLVMConstShl⚠ 	 [
Experimental
]

LLVMConstStringInContext⚠ 	 [
Experimental
]

LLVMConstStructInContext⚠ 	 [
Experimental
]

LLVMConstSub⚠ 	 [
Experimental
]

LLVMConstTrunc⚠ 	 [
Experimental
]

LLVMConstUDiv⚠ 	 [
Experimental
]

LLVMConstUIToFP⚠ 	 [
Experimental
]

LLVMConstURem⚠ 	 [
Experimental
]

LLVMConstVector⚠ 	 [
Experimental
]

LLVMConstXor⚠ 	 [
Experimental
]

LLVMConstZExt⚠ 	 [
Experimental
]

LLVMContextCreate⚠ 	 [
Experimental
]

LLVMContextDispose⚠ 	 [
Experimental
]

LLVMContextSetDiagnosticHandler⚠ 	 [
Experimental
]

LLVMCountParamTypes⚠ 	 [
Experimental
]

LLVMCountParams⚠ 	 [
Experimental
]

LLVMCountStructElementTypes⚠ 	 [
Experimental
]

LLVMCreateBuilderInContext⚠ 	 [
Experimental
]

LLVMCreateFunctionPassManagerForModule⚠ 	 [
Experimental
]
Creates a function-by-function pass manager

LLVMCreateObjectFile⚠ 	 [
Experimental
]
Opens an object file.

LLVMCreatePassManager⚠ 	 [
Experimental
]
Creates a pass manager.

LLVMCreateTargetData⚠ 	 [
Experimental
]
Creates target data from a target layout string.

LLVMDeleteBasicBlock⚠ 	 [
Experimental
]

LLVMDeleteGlobal⚠ 	 [
Experimental
]

LLVMDisposeBuilder⚠ 	 [
Experimental
]

LLVMDisposeModule⚠ 	 [
Experimental
]

LLVMDisposeObjectFile⚠ 	 [
Experimental
]
Closes an object file.

LLVMDisposePassManager⚠ 	 [
Experimental
]
Disposes a pass manager.

LLVMDisposeSectionIterator⚠ 	 [
Experimental
]
Destroys a section iterator.

LLVMDisposeTargetData⚠ 	 [
Experimental
]
Disposes target data.

LLVMDoubleTypeInContext⚠ 	 [
Experimental
]

LLVMDumpModule⚠ 	 [
Experimental
]
See Module::dump.

LLVMFloatTypeInContext⚠ 	 [
Experimental
]

LLVMFunctionType⚠ 	 [
Experimental
]

LLVMGetAlignment⚠ 	 [
Experimental
]

LLVMGetArrayLength⚠ 	 [
Experimental
]

LLVMGetBasicBlockParent⚠ 	 [
Experimental
]

LLVMGetCurrentDebugLocation⚠ 	 [
Experimental
]

LLVMGetDataLayout⚠ 	 [
Experimental
]
Data layout. See Module::getDataLayout.

LLVMGetElementType⚠ 	 [
Experimental
]

LLVMGetFirstBasicBlock⚠ 	 [
Experimental
]

LLVMGetFirstFunction⚠ 	 [
Experimental
]

LLVMGetFirstGlobal⚠ 	 [
Experimental
]

LLVMGetFirstInstruction⚠ 	 [
Experimental
]

LLVMGetFirstUse⚠ 	 [
Experimental
]

LLVMGetGlobalParent⚠ 	 [
Experimental
]

LLVMGetInitializer⚠ 	 [
Experimental
]

LLVMGetInsertBlock⚠ 	 [
Experimental
]

LLVMGetInstructionParent⚠ 	 [
Experimental
]

LLVMGetIntTypeWidth⚠ 	 [
Experimental
]

LLVMGetMDKindIDInContext⚠ 	 [
Experimental
]

LLVMGetModuleContext⚠ 	 [
Experimental
]

LLVMGetNamedFunction⚠ 	 [
Experimental
]

LLVMGetNamedGlobal⚠ 	 [
Experimental
]

LLVMGetNextFunction⚠ 	 [
Experimental
]

LLVMGetNextGlobal⚠ 	 [
Experimental
]

LLVMGetNextUse⚠ 	 [
Experimental
]

LLVMGetOperand⚠ 	 [
Experimental
]

LLVMGetParam⚠ 	 [
Experimental
]

LLVMGetParamTypes⚠ 	 [
Experimental
]

LLVMGetReturnType⚠ 	 [
Experimental
]

LLVMGetSection⚠ 	 [
Experimental
]

LLVMGetSectionContents⚠ 	 [
Experimental
]
Returns the current section contents as a string buffer.

LLVMGetSectionSize⚠ 	 [
Experimental
]
Returns the current section size.

LLVMGetSections⚠ 	 [
Experimental
]
Enumerates the sections in an object file.

LLVMGetStructElementTypes⚠ 	 [
Experimental
]

LLVMGetUndef⚠ 	 [
Experimental
]

LLVMGetUser⚠ 	 [
Experimental
]

LLVMGetValueName⚠ 	 [
Experimental
]

LLVMGetVectorSize⚠ 	 [
Experimental
]

LLVMInitializePasses⚠ 	 [
Experimental
]

LLVMInstructionEraseFromParent⚠ 	 [
Experimental
]

LLVMInt16TypeInContext⚠ 	 [
Experimental
]

LLVMInt1TypeInContext⚠ 	 [
Experimental
]

LLVMInt32TypeInContext⚠ 	 [
Experimental
]

LLVMInt64TypeInContext⚠ 	 [
Experimental
]

LLVMInt8TypeInContext⚠ 	 [
Experimental
]

LLVMIntTypeInContext⚠ 	 [
Experimental
]

LLVMIsAConstantInt⚠ 	 [
Experimental
]

LLVMIsAGlobalVariable⚠ 	 [
Experimental
]

LLVMIsAStoreInst⚠ 	 [
Experimental
]

LLVMIsATerminatorInst⚠ 	 [
Experimental
]

LLVMIsDeclaration⚠ 	 [
Experimental
]

LLVMIsGlobalConstant⚠ 	 [
Experimental
]

LLVMIsNull⚠ 	 [
Experimental
]

LLVMIsPackedStruct⚠ 	 [
Experimental
]

LLVMIsSectionIteratorAtEnd⚠ 	 [
Experimental
]
Returns true if the section iterator is at the end of the section list:

LLVMIsUndef⚠ 	 [
Experimental
]

LLVMMDNodeInContext⚠ 	 [
Experimental
]

LLVMMDStringInContext⚠ 	 [
Experimental
]

LLVMModuleCreateWithNameInContext⚠ 	 [
Experimental
]

LLVMMoveToNextSection⚠ 	 [
Experimental
]
Moves the section iterator to point to the next section.

LLVMOffsetOfElement⚠ 	 [
Experimental
]
Computes the byte offset of the indexed struct element for a target.

LLVMPassManagerBuilderCreate⚠ 	 [
Experimental
]

LLVMPassManagerBuilderDispose⚠ 	 [
Experimental
]

LLVMPassManagerBuilderPopulateFunctionPassManager⚠ 	 [
Experimental
]

LLVMPassManagerBuilderPopulateLTOPassManager⚠ 	 [
Experimental
]

LLVMPassManagerBuilderPopulateModulePassManager⚠ 	 [
Experimental
]

LLVMPassManagerBuilderSetDisableUnrollLoops⚠ 	 [
Experimental
]

LLVMPassManagerBuilderSetSizeLevel⚠ 	 [
Experimental
]

LLVMPassManagerBuilderUseInlinerWithThreshold⚠ 	 [
Experimental
]

LLVMPointerType⚠ 	 [
Experimental
]

LLVMPositionBuilder⚠ 	 [
Experimental
]

LLVMPositionBuilderAtEnd⚠ 	 [
Experimental
]

LLVMPositionBuilderBefore⚠ 	 [
Experimental
]

LLVMPreferredAlignmentOfType⚠ 	 [
Experimental
]
Returns the preferred alignment of a type.

LLVMReplaceAllUsesWith⚠ 	 [
Experimental
]

LLVMRunPassManager⚠ 	 [
Experimental
]
Runs a pass manager on a module.

LLVMRustAddAlwaysInlinePass⚠ 	 [
Experimental
]

LLVMRustAddAnalysisPasses⚠ 	 [
Experimental
]

LLVMRustAddBuilderLibraryInfo⚠ 	 [
Experimental
]

LLVMRustAddCallSiteAttribute⚠ 	 [
Experimental
]

LLVMRustAddDereferenceableAttr⚠ 	 [
Experimental
]

LLVMRustAddDereferenceableCallSiteAttr⚠ 	 [
Experimental
]

LLVMRustAddFunctionAttrStringValue⚠ 	 [
Experimental
]

LLVMRustAddFunctionAttribute⚠ 	 [
Experimental
]

LLVMRustAddHandler⚠ 	 [
Experimental
]

LLVMRustAddLibraryInfo⚠ 	 [
Experimental
]

LLVMRustAddModuleFlag⚠ 	 [
Experimental
]

LLVMRustAddPass⚠ 	 [
Experimental
]

LLVMRustAppendModuleInlineAsm⚠ 	 [
Experimental
]

LLVMRustArchiveChildData⚠ 	 [
Experimental
]

LLVMRustArchiveChildFree⚠ 	 [
Experimental
]

LLVMRustArchiveChildName⚠ 	 [
Experimental
]

LLVMRustArchiveIteratorFree⚠ 	 [
Experimental
]

LLVMRustArchiveIteratorNew⚠ 	 [
Experimental
]

LLVMRustArchiveIteratorNext⚠ 	 [
Experimental
]

LLVMRustArchiveMemberFree⚠ 	 [
Experimental
]

LLVMRustArchiveMemberNew⚠ 	 [
Experimental
]

LLVMRustArrayType⚠ 	 [
Experimental
]

LLVMRustBuildAtomicCmpXchg⚠ 	 [
Experimental
]

LLVMRustBuildAtomicFence⚠ 	 [
Experimental
]

LLVMRustBuildAtomicLoad⚠ 	 [
Experimental
]

LLVMRustBuildAtomicStore⚠ 	 [
Experimental
]

LLVMRustBuildCall⚠ 	 [
Experimental
]

LLVMRustBuildCatchPad⚠ 	 [
Experimental
]

LLVMRustBuildCatchRet⚠ 	 [
Experimental
]

LLVMRustBuildCatchSwitch⚠ 	 [
Experimental
]

LLVMRustBuildCleanupPad⚠ 	 [
Experimental
]

LLVMRustBuildCleanupRet⚠ 	 [
Experimental
]

LLVMRustBuildIntCast⚠ 	 [
Experimental
]

LLVMRustBuildInvoke⚠ 	 [
Experimental
]

LLVMRustBuildLandingPad⚠ 	 [
Experimental
]

LLVMRustBuildOperandBundleDef⚠ 	 [
Experimental
]

LLVMRustConfigurePassManagerBuilder⚠ 	 [
Experimental
]

LLVMRustConstInt128Get⚠ 	 [
Experimental
]

LLVMRustCreateMemoryBufferWithContentsOfFile⚠ 	 [
Experimental
]
Reads the given file and returns it as a memory buffer. Use LLVMDisposeMemoryBuffer() to get rid of it.

LLVMRustCreateTargetMachine⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreate⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateArrayType⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateBasicType⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateCompileUnit⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateDebugLocation⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateEnumerationType⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateEnumerator⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateFile⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateFunction⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateLexicalBlock⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateLexicalBlockFile⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateMemberType⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateNameSpace⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateOpDeref⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateOpPlus⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreatePointerType⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateStaticVariable⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateStructType⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateSubroutineType⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateTemplateTypeParameter⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateUnionType⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateVariable⚠ 	 [
Experimental
]

LLVMRustDIBuilderCreateVectorType⚠ 	 [
Experimental
]

LLVMRustDIBuilderDispose⚠ 	 [
Experimental
]

LLVMRustDIBuilderFinalize⚠ 	 [
Experimental
]

LLVMRustDIBuilderGetOrCreateArray⚠ 	 [
Experimental
]

LLVMRustDIBuilderGetOrCreateSubrange⚠ 	 [
Experimental
]

LLVMRustDIBuilderInsertDeclareAtEnd⚠ 	 [
Experimental
]

LLVMRustDICompositeTypeSetTypeArray⚠ 	 [
Experimental
]

LLVMRustDebugMetadataVersion⚠ 	 [
Experimental
]

LLVMRustDestroyArchive⚠ 	 [
Experimental
]

LLVMRustDisposeTargetMachine⚠ 	 [
Experimental
]

LLVMRustFindAndCreatePass⚠ 	 [
Experimental
]

LLVMRustFreeOperandBundleDef⚠ 	 [
Experimental
]

LLVMRustGetDiagInfoKind⚠ 	 [
Experimental
]

LLVMRustGetLastError⚠ 	 [
Experimental
]
Returns a string describing the last error caused by an LLVMRust* call.

LLVMRustGetLinkage⚠ 	 [
Experimental
]

LLVMRustGetModuleDataLayout⚠ 	 [
Experimental
]

LLVMRustGetNamedValue⚠ 	 [
Experimental
]

LLVMRustGetOrInsertFunction⚠ 	 [
Experimental
]

LLVMRustGetOrInsertGlobal⚠ 	 [
Experimental
]

LLVMRustGetSectionName⚠ 	 [
Experimental
]

LLVMRustGetTypeKind⚠ 	 [
Experimental
]
See llvm::LLVMTypeKind::getTypeID.

LLVMRustGetValueContext⚠ 	 [
Experimental
]
See llvm::Value::getContext

LLVMRustGetVisibility⚠ 	 [
Experimental
]

LLVMRustHasFeature⚠ 	 [
Experimental
]

LLVMRustInlineAsm⚠ 	 [
Experimental
]
Prepares inline assembly.

LLVMRustLinkInExternalBitcode⚠ 	 [
Experimental
]

LLVMRustMarkAllFunctionsNounwind⚠ 	 [
Experimental
]

LLVMRustMetadataAsValue⚠ 	 [
Experimental
]

LLVMRustMetadataTypeInContext⚠ 	 [
Experimental
]

LLVMRustOpenArchive⚠ 	 [
Experimental
]

LLVMRustPassKind⚠ 	 [
Experimental
]

LLVMRustPositionBuilderAtStart⚠ 	 [
Experimental
]

LLVMRustPrintModule⚠ 	 [
Experimental
]

LLVMRustPrintPassTimings⚠ 	 [
Experimental
]
Print the pass timings since static dtors aren't picking them up.

LLVMRustPrintPasses⚠ 	 [
Experimental
]

LLVMRustPrintTargetCPUs⚠ 	 [
Experimental
]

LLVMRustPrintTargetFeatures⚠ 	 [
Experimental
]

LLVMRustRemoveFunctionAttributes⚠ 	 [
Experimental
]

LLVMRustRunFunctionPassManager⚠ 	 [
Experimental
]

LLVMRustRunRestrictionPass⚠ 	 [
Experimental
]

LLVMRustSetComdat⚠ 	 [
Experimental
]

LLVMRustSetDataLayoutFromTargetMachine⚠ 	 [
Experimental
]

LLVMRustSetDebug⚠ 	 [
Experimental
]
Enables LLVM debug output.

LLVMRustSetHasUnsafeAlgebra⚠ 	 [
Experimental
]

LLVMRustSetInlineAsmDiagnosticHandler⚠ 	 [
Experimental
]

LLVMRustSetLLVMOptions⚠ 	 [
Experimental
]

LLVMRustSetLinkage⚠ 	 [
Experimental
]

LLVMRustSetModulePIELevel⚠ 	 [
Experimental
]

LLVMRustSetNormalizedTarget⚠ 	 [
Experimental
]

LLVMRustSetVisibility⚠ 	 [
Experimental
]

LLVMRustStringWriteImpl⚠ 	 [
Experimental
]
Appending to a Rust string -- used by RawRustStringOstream.

LLVMRustUnpackInlineAsmDiagnostic⚠ 	 [
Experimental
]

LLVMRustUnpackOptimizationDiagnostic⚠ 	 [
Experimental
]

LLVMRustUnsetComdat⚠ 	 [
Experimental
]

LLVMRustVersionMajor⚠ 	 [
Experimental
]

LLVMRustVersionMinor⚠ 	 [
Experimental
]

LLVMRustWriteArchive⚠ 	 [
Experimental
]

LLVMRustWriteDebugLocToString⚠ 	 [
Experimental
]

LLVMRustWriteDiagnosticInfoToString⚠ 	 [
Experimental
]

LLVMRustWriteOutputFile⚠ 	 [
Experimental
]

LLVMRustWriteSMDiagnosticToString⚠ 	 [
Experimental
]

LLVMRustWriteTwineToString⚠ 	 [
Experimental
]

LLVMRustWriteTypeToString⚠ 	 [
Experimental
]

LLVMRustWriteValueToString⚠ 	 [
Experimental
]

LLVMSetAlignment⚠ 	 [
Experimental
]

LLVMSetCleanup⚠ 	 [
Experimental
]

LLVMSetCurrentDebugLocation⚠ 	 [
Experimental
]

LLVMSetDLLStorageClass⚠ 	 [
Experimental
]

LLVMSetDataLayout⚠ 	 [
Experimental
]

LLVMSetFunctionCallConv⚠ 	 [
Experimental
]

LLVMSetGlobalConstant⚠ 	 [
Experimental
]

LLVMSetInitializer⚠ 	 [
Experimental
]

LLVMSetInstDebugLocation⚠ 	 [
Experimental
]

LLVMSetInstructionCallConv⚠ 	 [
Experimental
]

LLVMSetMetadata⚠ 	 [
Experimental
]

LLVMSetModuleInlineAsm⚠ 	 [
Experimental
]
See Module::setModuleInlineAsm.

LLVMSetPersonalityFn⚠ 	 [
Experimental
]

LLVMSetSection⚠ 	 [
Experimental
]

LLVMSetTailCall⚠ 	 [
Experimental
]

LLVMSetThreadLocal⚠ 	 [
Experimental
]

LLVMSetUnnamedAddr⚠ 	 [
Experimental
]

LLVMSetValueName⚠ 	 [
Experimental
]

LLVMSetVolatile⚠ 	 [
Experimental
]

LLVMSizeOf⚠ 	 [
Experimental
]

LLVMSizeOfTypeInBits⚠ 	 [
Experimental
]
Number of bytes clobbered when doing a Store to *T.

LLVMStartMultithreaded⚠ 	 [
Experimental
]

LLVMStructCreateNamed⚠ 	 [
Experimental
]

LLVMStructSetBody⚠ 	 [
Experimental
]

LLVMStructTypeInContext⚠ 	 [
Experimental
]

LLVMTypeOf⚠ 	 [
Experimental
]

LLVMVectorType⚠ 	 [
Experimental
]

LLVMVoidTypeInContext⚠ 	 [
Experimental
]

LLVMWriteBitcodeToFile⚠ 	 [
Experimental
]
Writes a module to the specified path. Returns 0 on success.

SetFunctionCallConv 	 [
Experimental
]

SetInstructionCallConv 	 [
Experimental
]

SetUniqueComdat 	 [
Experimental
]

SetUnnamedAddr 	 [
Experimental
]

UnsetComdat 	 [
Experimental
]

build_string 	 [
Experimental
]

debug_loc_to_string⚠ 	 [
Experimental
]

get_param 	 [
Experimental
]
Safe wrapper around LLVMGetParam, because segfaults are no fun.

initialize_available_targets 	 [
Experimental
]

last_error 	 [
Experimental
]

mk_section_iter 	 [
Experimental
]

set_thread_local 	 [
Experimental
]

twine_to_string⚠ 	 [
Experimental
]

Type Definitions

ArchiveChildRef 	 [
Experimental
]

ArchiveIteratorRef 	 [
Experimental
]

ArchiveRef 	 [
Experimental
]

BasicBlockRef 	 [
Experimental
]

Bool 	 [
Experimental
]

BuilderRef 	 [
Experimental
]

ContextRef 	 [
Experimental
]

DebugLocRef 	 [
Experimental
]

DiagnosticHandler 	 [
Experimental
]

DiagnosticInfoRef 	 [
Experimental
]

ExecutionEngineRef 	 [
Experimental
]

InlineAsmDiagHandler 	 [
Experimental
]

MemoryBufferRef 	 [
Experimental
]

MetadataRef 	 [
Experimental
]

ModuleRef 	 [
Experimental
]

ObjectFileRef 	 [
Experimental
]

Opcode 	 [
Experimental
]

OperandBundleDefRef 	 [
Experimental
]

PassManagerBuilderRef 	 [
Experimental
]

PassManagerRef 	 [
Experimental
]

PassRef 	 [
Experimental
]

RustArchiveMemberRef 	 [
Experimental
]

SMDiagnosticRef 	 [
Experimental
]

SectionIteratorRef 	 [
Experimental
]

TargetDataRef 	 [
Experimental
]

TargetMachineRef 	 [
Experimental
]

TwineRef 	 [
Experimental
]

TypeRef 	 [
Experimental
]

UseRef 	 [
Experimental
]

ValueRef 	 [
Experimental
]