Struct rustc_llvm::archive_ro::ArchiveRO

pub struct ArchiveRO { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl ArchiveRO

pub fn open(dst: &Path) -> Result<ArchiveRO, String>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Opens a static archive for read-only purposes. This is more optimized than the open method because it uses LLVM's internal Archive class rather than shelling out to ar for everything.

If this archive is used with a mutable method, then an error will be raised.

pub fn raw(&self) -> ArchiveRef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn iter(&self) -> Iter

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?