Module rustc_llvm::ffi

Unstable (rustc_private)

Modules

debuginfo 	 [Unstable]

Structs

Attribute 	 [Unstable]

Enums

ArchiveChild_opaque 	 [Unstable]

ArchiveIterator_opaque 	 [Unstable]

ArchiveKind 	 [Unstable]
LLVMRustArchiveKind

Archive_opaque 	 [Unstable]

AsmDialect 	 [Unstable]
LLVMRustAsmDialect

AtomicOrdering 	 [Unstable]
LLVMAtomicOrdering

AtomicRmwBinOp 	 [Unstable]
LLVMAtomicRmwBinOp

BasicBlock_opaque 	 [Unstable]

Builder_opaque 	 [Unstable]

CallConv 	 [Unstable]
LLVM CallingConv::ID. Should we wrap this?

CodeGenOptLevel 	 [Unstable]
LLVMRustCodeGenOptLevel

CodeModel 	 [Unstable]
LLVMRustCodeModel

Context_opaque 	 [Unstable]

DLLStorageClass 	 [Unstable]
LLVMDLLStorageClass

DebugLoc_opaque 	 [Unstable]

DiagnosticInfo_opaque 	 [Unstable]

DiagnosticKind 	 [Unstable]
LLVMRustDiagnosticKind

DiagnosticSeverity 	 [Unstable]
LLVMDiagnosticSeverity

ExecutionEngine_opaque 	 [Unstable]

FileType 	 [Unstable]
LLVMRustFileType

IntPredicate 	 [Unstable]
LLVMIntPredicate

LLVMRustResult 	 [Unstable]

Linkage 	 [Unstable]
LLVMLinkage

MemoryBuffer_opaque 	 [Unstable]

MetadataType 	 [Unstable]
Enum pinned in LLVMContext, used in LLVMSetMetadata so ABI-stable.

Metadata_opaque 	 [Unstable]

Module_opaque 	 [Unstable]

ObjectFile_opaque 	 [Unstable]

OperandBundleDef_opaque 	 [Unstable]

PassKind 	 [Unstable]
LLVMRustPassKind

PassManagerBuilder_opaque 	 [Unstable]

PassManager_opaque 	 [Unstable]

Pass_opaque 	 [Unstable]

RealPredicate 	 [Unstable]
LLVMRealPredicate

RelocMode 	 [Unstable]
LLVMRelocMode

RustArchiveMember_opaque 	 [Unstable]

SMDiagnostic_opaque 	 [Unstable]

SectionIterator_opaque 	 [Unstable]

SynchronizationScope 	 [Unstable]
LLVMRustSynchronizationScope

TargetData_opaque 	 [Unstable]

TargetMachine_opaque 	 [Unstable]

Twine_opaque 	 [Unstable]

TypeKind 	 [Unstable]
LLVMTypeKind

Type_opaque 	 [Unstable]

Use_opaque 	 [Unstable]

Value_opaque 	 [Unstable]

Constants

False 	 [Unstable]

True 	 [Unstable]

Functions

LLVMABIAlignmentOfType 	 [Unstable]
Returns the minimum alignment of a type.

LLVMABISizeOfType 	 [Unstable]
Distance between successive elements in an array of T. Includes ABI padding.

LLVMAddAggressiveDCEPass 	 [Unstable]

LLVMAddAlias 	 [Unstable]

LLVMAddArgumentPromotionPass 	 [Unstable]

LLVMAddAttribute 	 [Unstable]

LLVMAddBasicAliasAnalysisPass 	 [Unstable]

LLVMAddCFGSimplificationPass 	 [Unstable]

LLVMAddCase 	 [Unstable]

LLVMAddClause 	 [Unstable]

LLVMAddConstantMergePass 	 [Unstable]

LLVMAddConstantPropagationPass 	 [Unstable]

LLVMAddCorrelatedValuePropagationPass 	 [Unstable]

LLVMAddDeadArgEliminationPass 	 [Unstable]

LLVMAddDeadStoreEliminationPass 	 [Unstable]

LLVMAddDestination 	 [Unstable]

LLVMAddEarlyCSEPass 	 [Unstable]

LLVMAddFunction 	 [Unstable]

LLVMAddFunctionAttrsPass 	 [Unstable]

LLVMAddFunctionInliningPass 	 [Unstable]

LLVMAddGVNPass 	 [Unstable]

LLVMAddGlobal 	 [Unstable]

LLVMAddGlobalDCEPass 	 [Unstable]

LLVMAddGlobalInAddressSpace 	 [Unstable]

LLVMAddGlobalOptimizerPass 	 [Unstable]

LLVMAddIPSCCPPass 	 [Unstable]

LLVMAddIncoming 	 [Unstable]

LLVMAddIndVarSimplifyPass 	 [Unstable]

LLVMAddInstrAttribute 	 [Unstable]

LLVMAddInstructionCombiningPass 	 [Unstable]

LLVMAddJumpThreadingPass 	 [Unstable]

LLVMAddLICMPass 	 [Unstable]

LLVMAddLoopDeletionPass 	 [Unstable]

LLVMAddLoopIdiomPass 	 [Unstable]

LLVMAddLoopRotatePass 	 [Unstable]

LLVMAddLoopUnrollPass 	 [Unstable]

LLVMAddLoopUnswitchPass 	 [Unstable]

LLVMAddMemCpyOptPass 	 [Unstable]

LLVMAddNamedMetadataOperand 	 [Unstable]

LLVMAddPruneEHPass 	 [Unstable]

LLVMAddReassociatePass 	 [Unstable]

LLVMAddSCCPPass 	 [Unstable]

LLVMAddScalarReplAggregatesPass 	 [Unstable]

LLVMAddScalarReplAggregatesPassSSA 	 [Unstable]

LLVMAddSimplifyLibCallsPass 	 [Unstable]

LLVMAddStripDeadPrototypesPass 	 [Unstable]

LLVMAddTailCallEliminationPass 	 [Unstable]

LLVMAddTypeBasedAliasAnalysisPass 	 [Unstable]

LLVMAddVerifierPass 	 [Unstable]
Adds a verification pass.

LLVMAlignOf 	 [Unstable]

LLVMAppendBasicBlockInContext 	 [Unstable]

LLVMBasicBlockAsValue 	 [Unstable]

LLVMBlockAddress 	 [Unstable]

LLVMBuildAShr 	 [Unstable]

LLVMBuildAdd 	 [Unstable]

LLVMBuildAggregateRet 	 [Unstable]

LLVMBuildAlloca 	 [Unstable]

LLVMBuildAnd 	 [Unstable]

LLVMBuildAtomicRMW 	 [Unstable]

LLVMBuildBinOp 	 [Unstable]

LLVMBuildBitCast 	 [Unstable]

LLVMBuildBr 	 [Unstable]

LLVMBuildCast 	 [Unstable]

LLVMBuildCondBr 	 [Unstable]

LLVMBuildExactSDiv 	 [Unstable]

LLVMBuildExtractElement 	 [Unstable]

LLVMBuildExtractValue 	 [Unstable]

LLVMBuildFAdd 	 [Unstable]

LLVMBuildFCmp 	 [Unstable]

LLVMBuildFDiv 	 [Unstable]

LLVMBuildFMul 	 [Unstable]

LLVMBuildFNeg 	 [Unstable]

LLVMBuildFPCast 	 [Unstable]

LLVMBuildFPExt 	 [Unstable]

LLVMBuildFPToSI 	 [Unstable]

LLVMBuildFPToUI 	 [Unstable]

LLVMBuildFPTrunc 	 [Unstable]

LLVMBuildFRem 	 [Unstable]

LLVMBuildFSub 	 [Unstable]

LLVMBuildFree 	 [Unstable]

LLVMBuildGEP 	 [Unstable]

LLVMBuildGlobalString 	 [Unstable]

LLVMBuildGlobalStringPtr 	 [Unstable]

LLVMBuildICmp 	 [Unstable]

LLVMBuildInBoundsGEP 	 [Unstable]

LLVMBuildIndirectBr 	 [Unstable]

LLVMBuildInsertElement 	 [Unstable]

LLVMBuildInsertValue 	 [Unstable]

LLVMBuildIntCast 	 [Unstable]

LLVMBuildIntToPtr 	 [Unstable]

LLVMBuildIsNotNull 	 [Unstable]

LLVMBuildIsNull 	 [Unstable]

LLVMBuildLShr 	 [Unstable]

LLVMBuildLoad 	 [Unstable]

LLVMBuildMul 	 [Unstable]

LLVMBuildNSWAdd 	 [Unstable]

LLVMBuildNSWMul 	 [Unstable]

LLVMBuildNSWNeg 	 [Unstable]

LLVMBuildNSWSub 	 [Unstable]

LLVMBuildNUWAdd 	 [Unstable]

LLVMBuildNUWMul 	 [Unstable]

LLVMBuildNUWNeg 	 [Unstable]

LLVMBuildNUWSub 	 [Unstable]

LLVMBuildNeg 	 [Unstable]

LLVMBuildNot 	 [Unstable]

LLVMBuildOr 	 [Unstable]

LLVMBuildPhi 	 [Unstable]

LLVMBuildPointerCast 	 [Unstable]

LLVMBuildPtrDiff 	 [Unstable]

LLVMBuildPtrToInt 	 [Unstable]

LLVMBuildResume 	 [Unstable]

LLVMBuildRet 	 [Unstable]

LLVMBuildRetVoid 	 [Unstable]

LLVMBuildSDiv 	 [Unstable]

LLVMBuildSExt 	 [Unstable]

LLVMBuildSExtOrBitCast 	 [Unstable]

LLVMBuildSIToFP 	 [Unstable]

LLVMBuildSRem 	 [Unstable]

LLVMBuildSelect 	 [Unstable]

LLVMBuildShl 	 [Unstable]

LLVMBuildShuffleVector 	 [Unstable]

LLVMBuildStore 	 [Unstable]

LLVMBuildStructGEP 	 [Unstable]

LLVMBuildSub 	 [Unstable]

LLVMBuildSwitch 	 [Unstable]

LLVMBuildTrunc 	 [Unstable]

LLVMBuildTruncOrBitCast 	 [Unstable]

LLVMBuildUDiv 	 [Unstable]

LLVMBuildUIToFP 	 [Unstable]

LLVMBuildURem 	 [Unstable]

LLVMBuildUnreachable 	 [Unstable]

LLVMBuildVAArg 	 [Unstable]

LLVMBuildXor 	 [Unstable]

LLVMBuildZExt 	 [Unstable]

LLVMBuildZExtOrBitCast 	 [Unstable]

LLVMCallFrameAlignmentOfType 	 [Unstable]
Returns the minimum alignment of a type when part of a call frame.

LLVMClearInsertionPosition 	 [Unstable]

LLVMCloneModule 	 [Unstable]

LLVMConstAShr 	 [Unstable]

LLVMConstAdd 	 [Unstable]

LLVMConstAllOnes 	 [Unstable]

LLVMConstAnd 	 [Unstable]

LLVMConstArray 	 [Unstable]

LLVMConstBitCast 	 [Unstable]

LLVMConstExactSDiv 	 [Unstable]

LLVMConstExtractElement 	 [Unstable]

LLVMConstExtractValue 	 [Unstable]

LLVMConstFAdd 	 [Unstable]

LLVMConstFCmp 	 [Unstable]

LLVMConstFDiv 	 [Unstable]

LLVMConstFMul 	 [Unstable]

LLVMConstFNeg 	 [Unstable]

LLVMConstFPCast 	 [Unstable]

LLVMConstFPExt 	 [Unstable]

LLVMConstFPToSI 	 [Unstable]

LLVMConstFPToUI 	 [Unstable]

LLVMConstFPTrunc 	 [Unstable]

LLVMConstFRem 	 [Unstable]

LLVMConstFSub 	 [Unstable]

LLVMConstGEP 	 [Unstable]

LLVMConstICmp 	 [Unstable]

LLVMConstInBoundsGEP 	 [Unstable]

LLVMConstInlineAsm 	 [Unstable]

LLVMConstInsertElement 	 [Unstable]

LLVMConstInsertValue 	 [Unstable]

LLVMConstInt 	 [Unstable]

LLVMConstIntCast 	 [Unstable]

LLVMConstIntGetSExtValue 	 [Unstable]

LLVMConstIntGetZExtValue 	 [Unstable]

LLVMConstIntOfString 	 [Unstable]

LLVMConstIntOfStringAndSize 	 [Unstable]

LLVMConstIntToPtr 	 [Unstable]

LLVMConstLShr 	 [Unstable]

LLVMConstMul 	 [Unstable]

LLVMConstNSWAdd 	 [Unstable]

LLVMConstNSWMul 	 [Unstable]

LLVMConstNSWNeg 	 [Unstable]

LLVMConstNSWSub 	 [Unstable]

LLVMConstNUWAdd 	 [Unstable]

LLVMConstNUWMul 	 [Unstable]

LLVMConstNUWNeg 	 [Unstable]

LLVMConstNUWSub 	 [Unstable]

LLVMConstNamedStruct 	 [Unstable]

LLVMConstNeg 	 [Unstable]

LLVMConstNot 	 [Unstable]

LLVMConstNull 	 [Unstable]

LLVMConstOr 	 [Unstable]

LLVMConstPointerCast 	 [Unstable]

LLVMConstPointerNull 	 [Unstable]

LLVMConstPtrToInt 	 [Unstable]

LLVMConstReal 	 [Unstable]

LLVMConstRealOfString 	 [Unstable]

LLVMConstRealOfStringAndSize 	 [Unstable]

LLVMConstSDiv 	 [Unstable]

LLVMConstSExt 	 [Unstable]

LLVMConstSExtOrBitCast 	 [Unstable]

LLVMConstSIToFP 	 [Unstable]

LLVMConstSRem 	 [Unstable]

LLVMConstSelect 	 [Unstable]

LLVMConstShl 	 [Unstable]

LLVMConstShuffleVector 	 [Unstable]

LLVMConstStringInContext 	 [Unstable]

LLVMConstStructInContext 	 [Unstable]

LLVMConstSub 	 [Unstable]

LLVMConstTrunc 	 [Unstable]

LLVMConstTruncOrBitCast 	 [Unstable]

LLVMConstUDiv 	 [Unstable]

LLVMConstUIToFP 	 [Unstable]

LLVMConstURem 	 [Unstable]

LLVMConstVector 	 [Unstable]

LLVMConstXor 	 [Unstable]

LLVMConstZExt 	 [Unstable]

LLVMConstZExtOrBitCast 	 [Unstable]

LLVMContextCreate 	 [Unstable]

LLVMContextDispose 	 [Unstable]

LLVMContextSetDiagnosticHandler 	 [Unstable]

LLVMCountBasicBlocks 	 [Unstable]

LLVMCountIncoming 	 [Unstable]

LLVMCountParamTypes 	 [Unstable]

LLVMCountParams 	 [Unstable]

LLVMCountStructElementTypes 	 [Unstable]

LLVMCreateBuilderInContext 	 [Unstable]

LLVMCreateFunctionPassManagerForModule 	 [Unstable]
Creates a function-by-function pass manager

LLVMCreateMemoryBufferWithMemoryRange 	 [Unstable]
Borrows the contents of the memory buffer (doesn't copy it)

LLVMCreateMemoryBufferWithMemoryRangeCopy 	 [Unstable]

LLVMCreateObjectFile 	 [Unstable]
Opens an object file.

LLVMCreatePassManager 	 [Unstable]
Creates a pass manager.

LLVMCreateTargetData 	 [Unstable]
Creates target data from a target layout string.

LLVMDeleteBasicBlock 	 [Unstable]

LLVMDeleteFunction 	 [Unstable]

LLVMDeleteGlobal 	 [Unstable]

LLVMDisposeBuilder 	 [Unstable]

LLVMDisposeMemoryBuffer 	 [Unstable]
Destroys a memory buffer.

LLVMDisposeModule 	 [Unstable]

LLVMDisposeObjectFile 	 [Unstable]
Closes an object file.

LLVMDisposePassManager 	 [Unstable]
Disposes a pass manager.

LLVMDisposeSectionIterator 	 [Unstable]
Destroys a section iterator.

LLVMDisposeTargetData 	 [Unstable]
Disposes target data.

LLVMDoubleTypeInContext 	 [Unstable]

LLVMDumpModule 	 [Unstable]
See Module::dump.

LLVMDumpValue 	 [Unstable]

LLVMFP128TypeInContext 	 [Unstable]

LLVMFinalizeFunctionPassManager 	 [Unstable]
Finalizes all the function passes scheduled in the manager

LLVMFloatTypeInContext 	 [Unstable]

LLVMFunctionType 	 [Unstable]

LLVMGetAlignment 	 [Unstable]

LLVMGetArrayLength 	 [Unstable]

LLVMGetAttribute 	 [Unstable]

LLVMGetBasicBlockParent 	 [Unstable]

LLVMGetBasicBlocks 	 [Unstable]

LLVMGetCurrentDebugLocation 	 [Unstable]

LLVMGetDataLayout 	 [Unstable]
Data layout. See Module::getDataLayout.

LLVMGetDiagInfoSeverity 	 [Unstable]

LLVMGetElementType 	 [Unstable]

LLVMGetEntryBasicBlock 	 [Unstable]

LLVMGetFirstBasicBlock 	 [Unstable]

LLVMGetFirstFunction 	 [Unstable]

LLVMGetFirstGlobal 	 [Unstable]

LLVMGetFirstInstruction 	 [Unstable]

LLVMGetFirstParam 	 [Unstable]

LLVMGetFirstUse 	 [Unstable]

LLVMGetFunctionAttr 	 [Unstable]

LLVMGetFunctionCallConv 	 [Unstable]

LLVMGetGC 	 [Unstable]

LLVMGetGlobalParent 	 [Unstable]

LLVMGetIncomingBlock 	 [Unstable]

LLVMGetIncomingValue 	 [Unstable]

LLVMGetInitializer 	 [Unstable]

LLVMGetInsertBlock 	 [Unstable]

LLVMGetInstructionCallConv 	 [Unstable]

LLVMGetInstructionParent 	 [Unstable]

LLVMGetIntTypeWidth 	 [Unstable]

LLVMGetIntrinsicID 	 [Unstable]

LLVMGetLastBasicBlock 	 [Unstable]

LLVMGetLastFunction 	 [Unstable]

LLVMGetLastGlobal 	 [Unstable]

LLVMGetLastInstruction 	 [Unstable]

LLVMGetLastParam 	 [Unstable]

LLVMGetLinkage 	 [Unstable]

LLVMGetMDKindIDInContext 	 [Unstable]

LLVMGetModuleContext 	 [Unstable]

LLVMGetNamedFunction 	 [Unstable]

LLVMGetNamedGlobal 	 [Unstable]

LLVMGetNextBasicBlock 	 [Unstable]

LLVMGetNextFunction 	 [Unstable]

LLVMGetNextGlobal 	 [Unstable]

LLVMGetNextInstruction 	 [Unstable]

LLVMGetNextParam 	 [Unstable]

LLVMGetNextUse 	 [Unstable]

LLVMGetNumOperands 	 [Unstable]

LLVMGetOperand 	 [Unstable]

LLVMGetParam 	 [Unstable]

LLVMGetParamParent 	 [Unstable]

LLVMGetParamTypes 	 [Unstable]

LLVMGetParams 	 [Unstable]

LLVMGetPointerAddressSpace 	 [Unstable]

LLVMGetPointerToGlobal 	 [Unstable]

LLVMGetPreviousBasicBlock 	 [Unstable]

LLVMGetPreviousFunction 	 [Unstable]

LLVMGetPreviousGlobal 	 [Unstable]

LLVMGetPreviousInstruction 	 [Unstable]

LLVMGetPreviousParam 	 [Unstable]

LLVMGetReturnType 	 [Unstable]

LLVMGetSection 	 [Unstable]

LLVMGetSectionContents 	 [Unstable]
Returns the current section contents as a string buffer.

LLVMGetSectionSize 	 [Unstable]
Returns the current section size.

LLVMGetSections 	 [Unstable]
Enumerates the sections in an object file.

LLVMGetStructElementTypes 	 [Unstable]

LLVMGetTarget 	 [Unstable]
Target triple. See Module::getTargetTriple.

LLVMGetTypeContext 	 [Unstable]
See llvm::LLVMType::getContext.

LLVMGetUndef 	 [Unstable]

LLVMGetUsedValue 	 [Unstable]

LLVMGetUser 	 [Unstable]

LLVMGetValueName 	 [Unstable]

LLVMGetVectorSize 	 [Unstable]

LLVMGetVisibility 	 [Unstable]

LLVMGetVolatile 	 [Unstable]

LLVMInitializeFunctionPassManager 	 [Unstable]
Initializes all the function passes scheduled in the manager

LLVMInitializePasses 	 [Unstable]

LLVMInsertBasicBlockInContext 	 [Unstable]

LLVMInsertIntoBuilder 	 [Unstable]

LLVMInsertIntoBuilderWithName 	 [Unstable]

LLVMInstructionEraseFromParent 	 [Unstable]

LLVMInt16TypeInContext 	 [Unstable]

LLVMInt1TypeInContext 	 [Unstable]

LLVMInt32TypeInContext 	 [Unstable]

LLVMInt64TypeInContext 	 [Unstable]

LLVMInt8TypeInContext 	 [Unstable]

LLVMIntTypeInContext 	 [Unstable]

LLVMIsAAllocaInst 	 [Unstable]

LLVMIsAArgument 	 [Unstable]

LLVMIsAConstantInt 	 [Unstable]

LLVMIsAGlobalVariable 	 [Unstable]

LLVMIsAStoreInst 	 [Unstable]

LLVMIsATerminatorInst 	 [Unstable]

LLVMIsConstant 	 [Unstable]

LLVMIsDeclaration 	 [Unstable]

LLVMIsFunctionVarArg 	 [Unstable]

LLVMIsGlobalConstant 	 [Unstable]

LLVMIsMultithreaded 	 [Unstable]

LLVMIsNull 	 [Unstable]

LLVMIsPackedStruct 	 [Unstable]

LLVMIsSectionIteratorAtEnd 	 [Unstable]
Returns true if the section iterator is at the end of the section list:

LLVMIsTailCall 	 [Unstable]

LLVMIsThreadLocal 	 [Unstable]

LLVMIsUndef 	 [Unstable]

LLVMLabelTypeInContext 	 [Unstable]

LLVMMDNodeInContext 	 [Unstable]

LLVMMDStringInContext 	 [Unstable]

LLVMModuleCreateWithNameInContext 	 [Unstable]

LLVMMoveBasicBlockAfter 	 [Unstable]

LLVMMoveBasicBlockBefore 	 [Unstable]

LLVMMoveToNextSection 	 [Unstable]
Moves the section iterator to point to the next section.

LLVMOffsetOfElement 	 [Unstable]
Computes the byte offset of the indexed struct element for a target.

LLVMPPCFP128TypeInContext 	 [Unstable]

LLVMPassManagerBuilderCreate 	 [Unstable]

LLVMPassManagerBuilderDispose 	 [Unstable]

LLVMPassManagerBuilderPopulateFunctionPassManager 	 [Unstable]

LLVMPassManagerBuilderPopulateLTOPassManager 	 [Unstable]

LLVMPassManagerBuilderPopulateModulePassManager 	 [Unstable]

LLVMPassManagerBuilderSetDisableSimplifyLibCalls 	 [Unstable]

LLVMPassManagerBuilderSetDisableUnitAtATime 	 [Unstable]

LLVMPassManagerBuilderSetDisableUnrollLoops 	 [Unstable]

LLVMPassManagerBuilderSetOptLevel 	 [Unstable]

LLVMPassManagerBuilderSetSizeLevel 	 [Unstable]

LLVMPassManagerBuilderUseInlinerWithThreshold 	 [Unstable]

LLVMPointerType 	 [Unstable]

LLVMPositionBuilder 	 [Unstable]

LLVMPositionBuilderAtEnd 	 [Unstable]

LLVMPositionBuilderBefore 	 [Unstable]

LLVMPreferredAlignmentOfType 	 [Unstable]
Returns the preferred alignment of a type.

LLVMRemoveAttribute 	 [Unstable]

LLVMRemoveFunctionAttr 	 [Unstable]

LLVMRemoveInstrAttribute 	 [Unstable]

LLVMReplaceAllUsesWith 	 [Unstable]

LLVMRunFunctionPassManager 	 [Unstable]
Runs the function passes on the provided function.

LLVMRunPassManager 	 [Unstable]
Runs a pass manager on a module.

LLVMRustAddAlwaysInlinePass 	 [Unstable]

LLVMRustAddAnalysisPasses 	 [Unstable]

LLVMRustAddBuilderLibraryInfo 	 [Unstable]

LLVMRustAddCallSiteAttribute 	 [Unstable]

LLVMRustAddDereferenceableAttr 	 [Unstable]

LLVMRustAddDereferenceableCallSiteAttr 	 [Unstable]

LLVMRustAddFunctionAttrString 	 [Unstable]

LLVMRustAddFunctionAttrStringValue 	 [Unstable]

LLVMRustAddFunctionAttribute 	 [Unstable]

LLVMRustAddHandler 	 [Unstable]

LLVMRustAddLibraryInfo 	 [Unstable]

LLVMRustAddModuleFlag 	 [Unstable]

LLVMRustAddPass 	 [Unstable]

LLVMRustArchiveChildData 	 [Unstable]

LLVMRustArchiveChildFree 	 [Unstable]

LLVMRustArchiveChildName 	 [Unstable]

LLVMRustArchiveIteratorFree 	 [Unstable]

LLVMRustArchiveIteratorNew 	 [Unstable]

LLVMRustArchiveIteratorNext 	 [Unstable]

LLVMRustArchiveMemberFree 	 [Unstable]

LLVMRustArchiveMemberNew 	 [Unstable]

LLVMRustArrayType 	 [Unstable]

LLVMRustBuildAtomicCmpXchg 	 [Unstable]

LLVMRustBuildAtomicFence 	 [Unstable]

LLVMRustBuildAtomicLoad 	 [Unstable]

LLVMRustBuildAtomicStore 	 [Unstable]

LLVMRustBuildCall 	 [Unstable]

LLVMRustBuildCatchPad 	 [Unstable]

LLVMRustBuildCatchRet 	 [Unstable]

LLVMRustBuildCatchSwitch 	 [Unstable]

LLVMRustBuildCleanupPad 	 [Unstable]

LLVMRustBuildCleanupRet 	 [Unstable]

LLVMRustBuildInvoke 	 [Unstable]

LLVMRustBuildLandingPad 	 [Unstable]

LLVMRustBuildOperandBundleDef 	 [Unstable]

LLVMRustConfigurePassManagerBuilder 	 [Unstable]

LLVMRustCreateMemoryBufferWithContentsOfFile 	 [Unstable]
Reads the given file and returns it as a memory buffer. Use LLVMDisposeMemoryBuffer() to get rid of it.

LLVMRustCreateTargetMachine 	 [Unstable]

LLVMRustDIBuilderCreate 	 [Unstable]

LLVMRustDIBuilderCreateArrayType 	 [Unstable]

LLVMRustDIBuilderCreateBasicType 	 [Unstable]

LLVMRustDIBuilderCreateCompileUnit 	 [Unstable]

LLVMRustDIBuilderCreateDebugLocation 	 [Unstable]

LLVMRustDIBuilderCreateEnumerationType 	 [Unstable]

LLVMRustDIBuilderCreateEnumerator 	 [Unstable]

LLVMRustDIBuilderCreateFile 	 [Unstable]

LLVMRustDIBuilderCreateFunction 	 [Unstable]

LLVMRustDIBuilderCreateLexicalBlock 	 [Unstable]

LLVMRustDIBuilderCreateMemberType 	 [Unstable]

LLVMRustDIBuilderCreateNameSpace 	 [Unstable]

LLVMRustDIBuilderCreateOpDeref 	 [Unstable]

LLVMRustDIBuilderCreateOpPlus 	 [Unstable]

LLVMRustDIBuilderCreatePointerType 	 [Unstable]

LLVMRustDIBuilderCreateStaticVariable 	 [Unstable]

LLVMRustDIBuilderCreateStructType 	 [Unstable]

LLVMRustDIBuilderCreateSubroutineType 	 [Unstable]

LLVMRustDIBuilderCreateTemplateTypeParameter 	 [Unstable]

LLVMRustDIBuilderCreateUnionType 	 [Unstable]

LLVMRustDIBuilderCreateVariable 	 [Unstable]

LLVMRustDIBuilderCreateVectorType 	 [Unstable]

LLVMRustDIBuilderDispose 	 [Unstable]

LLVMRustDIBuilderFinalize 	 [Unstable]

LLVMRustDIBuilderGetOrCreateArray 	 [Unstable]

LLVMRustDIBuilderGetOrCreateSubrange 	 [Unstable]

LLVMRustDIBuilderInsertDeclareAtEnd 	 [Unstable]

LLVMRustDIBuilderInsertDeclareBefore 	 [Unstable]

LLVMRustDICompositeTypeSetTypeArray 	 [Unstable]

LLVMRustDebugMetadataVersion 	 [Unstable]

LLVMRustDestroyArchive 	 [Unstable]

LLVMRustDisposeTargetMachine 	 [Unstable]

LLVMRustFindAndCreatePass 	 [Unstable]

LLVMRustFreeOperandBundleDef 	 [Unstable]

LLVMRustGetDiagInfoKind 	 [Unstable]

LLVMRustGetLastError 	 [Unstable]
Returns a string describing the last error caused by an LLVMRust* call.

LLVMRustGetModuleDataLayout 	 [Unstable]

LLVMRustGetNamedValue 	 [Unstable]

LLVMRustGetOrInsertFunction 	 [Unstable]

LLVMRustGetOrInsertGlobal 	 [Unstable]

LLVMRustGetSectionName 	 [Unstable]

LLVMRustGetTypeKind 	 [Unstable]
See llvm::LLVMTypeKind::getTypeID.

LLVMRustHasFeature 	 [Unstable]

LLVMRustInlineAsm 	 [Unstable]
Prepares inline assembly.

LLVMRustLinkInExternalBitcode 	 [Unstable]

LLVMRustMarkAllFunctionsNounwind 	 [Unstable]

LLVMRustMetadataTypeInContext 	 [Unstable]

LLVMRustOpenArchive 	 [Unstable]

LLVMRustPassKind 	 [Unstable]

LLVMRustPositionBuilderAtStart 	 [Unstable]

LLVMRustPrintModule 	 [Unstable]

LLVMRustPrintPassTimings 	 [Unstable]
Print the pass timings since static dtors aren't picking them up.

LLVMRustPrintPasses 	 [Unstable]

LLVMRustPrintTargetCPUs 	 [Unstable]

LLVMRustPrintTargetFeatures 	 [Unstable]

LLVMRustRemoveFunctionAttrString 	 [Unstable]

LLVMRustRemoveFunctionAttributes 	 [Unstable]

LLVMRustRunFunctionPassManager 	 [Unstable]

LLVMRustRunRestrictionPass 	 [Unstable]

LLVMRustSetComdat 	 [Unstable]

LLVMRustSetDataLayoutFromTargetMachine 	 [Unstable]

LLVMRustSetDebug 	 [Unstable]
Enables LLVM debug output.

LLVMRustSetHasUnsafeAlgebra 	 [Unstable]

LLVMRustSetInlineAsmDiagnosticHandler 	 [Unstable]

LLVMRustSetLLVMOptions 	 [Unstable]

LLVMRustSetModulePIELevel 	 [Unstable]

LLVMRustSetNormalizedTarget 	 [Unstable]

LLVMRustSetPersonalityFn 	 [Unstable]

LLVMRustUnpackInlineAsmDiagnostic 	 [Unstable]

LLVMRustUnpackOptimizationDiagnostic 	 [Unstable]

LLVMRustUnsetComdat 	 [Unstable]

LLVMRustVersionMajor 	 [Unstable]

LLVMRustVersionMinor 	 [Unstable]

LLVMRustWriteArchive 	 [Unstable]

LLVMRustWriteDebugLocToString 	 [Unstable]

LLVMRustWriteDiagnosticInfoToString 	 [Unstable]

LLVMRustWriteOutputFile 	 [Unstable]

LLVMRustWriteSMDiagnosticToString 	 [Unstable]

LLVMRustWriteTwineToString 	 [Unstable]

LLVMRustWriteTypeToString 	 [Unstable]

LLVMRustWriteValueToString 	 [Unstable]

LLVMSetAlignment 	 [Unstable]

LLVMSetCleanup 	 [Unstable]

LLVMSetCurrentDebugLocation 	 [Unstable]

LLVMSetDLLStorageClass 	 [Unstable]

LLVMSetDataLayout 	 [Unstable]

LLVMSetFunctionCallConv 	 [Unstable]

LLVMSetGC 	 [Unstable]

LLVMSetGlobalConstant 	 [Unstable]

LLVMSetInitializer 	 [Unstable]

LLVMSetInstDebugLocation 	 [Unstable]

LLVMSetInstrParamAlignment 	 [Unstable]

LLVMSetInstructionCallConv 	 [Unstable]

LLVMSetLinkage 	 [Unstable]

LLVMSetMetadata 	 [Unstable]

LLVMSetModuleInlineAsm 	 [Unstable]
See Module::setModuleInlineAsm.

LLVMSetOperand 	 [Unstable]

LLVMSetParamAlignment 	 [Unstable]

LLVMSetSection 	 [Unstable]

LLVMSetTailCall 	 [Unstable]

LLVMSetTarget 	 [Unstable]

LLVMSetThreadLocal 	 [Unstable]

LLVMSetUnnamedAddr 	 [Unstable]

LLVMSetValueName 	 [Unstable]

LLVMSetVisibility 	 [Unstable]

LLVMSetVolatile 	 [Unstable]

LLVMSizeOf 	 [Unstable]

LLVMSizeOfTypeInBits 	 [Unstable]
Number of bytes clobbered when doing a Store to *T.

LLVMStartMultithreaded 	 [Unstable]

LLVMStoreSizeOfType 	 [Unstable]
Number of bytes clobbered when doing a Store to *T.

LLVMStructCreateNamed 	 [Unstable]

LLVMStructSetBody 	 [Unstable]

LLVMStructTypeInContext 	 [Unstable]

LLVMTypeOf 	 [Unstable]

LLVMValueAsBasicBlock 	 [Unstable]

LLVMValueIsBasicBlock 	 [Unstable]

LLVMVectorType 	 [Unstable]

LLVMVoidTypeInContext 	 [Unstable]

LLVMWriteBitcodeToFile 	 [Unstable]
Writes a module to the specified path. Returns 0 on success.

LLVMX86FP80TypeInContext 	 [Unstable]

Type Definitions

ArchiveChildRef 	 [Unstable]

ArchiveIteratorRef 	 [Unstable]

ArchiveRef 	 [Unstable]

BasicBlockRef 	 [Unstable]

Bool 	 [Unstable]

BuilderRef 	 [Unstable]

ContextRef 	 [Unstable]

DebugLocRef 	 [Unstable]

DiagnosticHandler 	 [Unstable]

DiagnosticInfoRef 	 [Unstable]

ExecutionEngineRef 	 [Unstable]

InlineAsmDiagHandler 	 [Unstable]

MemoryBufferRef 	 [Unstable]

MetadataRef 	 [Unstable]

ModuleRef 	 [Unstable]

ObjectFileRef 	 [Unstable]

Opcode 	 [Unstable]

OperandBundleDefRef 	 [Unstable]

PassManagerBuilderRef 	 [Unstable]

PassManagerRef 	 [Unstable]

PassRef 	 [Unstable]

RustArchiveMemberRef 	 [Unstable]

SMDiagnosticRef 	 [Unstable]

SectionIteratorRef 	 [Unstable]

TargetDataRef 	 [Unstable]

TargetMachineRef 	 [Unstable]

TwineRef 	 [Unstable]

TypeRef 	 [Unstable]

UseRef 	 [Unstable]

ValueRef 	 [Unstable]