Struct rustc_llvm::Attribute

pub struct Attribute {
    // some fields omitted
}

Unstable (rustc_private)

Methods

impl Attribute

const ZExt: Attribute = Attribute{bits: 1 << 0,}

Unstable (rustc_private)

const SExt: Attribute = Attribute{bits: 1 << 1,}

Unstable (rustc_private)

const NoReturn: Attribute = Attribute{bits: 1 << 2,}

Unstable (rustc_private)

const InReg: Attribute = Attribute{bits: 1 << 3,}

Unstable (rustc_private)

const StructRet: Attribute = Attribute{bits: 1 << 4,}

Unstable (rustc_private)

const NoUnwind: Attribute = Attribute{bits: 1 << 5,}

Unstable (rustc_private)

const NoAlias: Attribute = Attribute{bits: 1 << 6,}

Unstable (rustc_private)

const ByVal: Attribute = Attribute{bits: 1 << 7,}

Unstable (rustc_private)

const Nest: Attribute = Attribute{bits: 1 << 8,}

Unstable (rustc_private)

const ReadNone: Attribute = Attribute{bits: 1 << 9,}

Unstable (rustc_private)

const ReadOnly: Attribute = Attribute{bits: 1 << 10,}

Unstable (rustc_private)

const NoInline: Attribute = Attribute{bits: 1 << 11,}

Unstable (rustc_private)

const AlwaysInline: Attribute = Attribute{bits: 1 << 12,}

Unstable (rustc_private)

const OptimizeForSize: Attribute = Attribute{bits: 1 << 13,}

Unstable (rustc_private)

const StackProtect: Attribute = Attribute{bits: 1 << 14,}

Unstable (rustc_private)

const StackProtectReq: Attribute = Attribute{bits: 1 << 15,}

Unstable (rustc_private)

const NoCapture: Attribute = Attribute{bits: 1 << 21,}

Unstable (rustc_private)

const NoRedZone: Attribute = Attribute{bits: 1 << 22,}

Unstable (rustc_private)

const NoImplicitFloat: Attribute = Attribute{bits: 1 << 23,}

Unstable (rustc_private)

const Naked: Attribute = Attribute{bits: 1 << 24,}

Unstable (rustc_private)

const InlineHint: Attribute = Attribute{bits: 1 << 25,}

Unstable (rustc_private)

const ReturnsTwice: Attribute = Attribute{bits: 1 << 29,}

Unstable (rustc_private)

const UWTable: Attribute = Attribute{bits: 1 << 30,}

Unstable (rustc_private)

const NonLazyBind: Attribute = Attribute{bits: 1 << 31,}

Unstable (rustc_private)

const SanitizeAddress: Attribute = Attribute{bits: 1 << 32,}

Unstable (rustc_private)

const MinSize: Attribute = Attribute{bits: 1 << 33,}

Unstable (rustc_private)

const NoDuplicate: Attribute = Attribute{bits: 1 << 34,}

Unstable (rustc_private)

const StackProtectStrong: Attribute = Attribute{bits: 1 << 35,}

Unstable (rustc_private)

const SanitizeThread: Attribute = Attribute{bits: 1 << 36,}

Unstable (rustc_private)

const SanitizeMemory: Attribute = Attribute{bits: 1 << 37,}

Unstable (rustc_private)

const NoBuiltin: Attribute = Attribute{bits: 1 << 38,}

Unstable (rustc_private)

const Returned: Attribute = Attribute{bits: 1 << 39,}

Unstable (rustc_private)

const Cold: Attribute = Attribute{bits: 1 << 40,}

Unstable (rustc_private)

const Builtin: Attribute = Attribute{bits: 1 << 41,}

Unstable (rustc_private)

const OptimizeNone: Attribute = Attribute{bits: 1 << 42,}

Unstable (rustc_private)

const InAlloca: Attribute = Attribute{bits: 1 << 43,}

Unstable (rustc_private)

const NonNull: Attribute = Attribute{bits: 1 << 44,}

Unstable (rustc_private)

const JumpTable: Attribute = Attribute{bits: 1 << 45,}

Unstable (rustc_private)

const Convergent: Attribute = Attribute{bits: 1 << 46,}

Unstable (rustc_private)

const SafeStack: Attribute = Attribute{bits: 1 << 47,}

Unstable (rustc_private)

const NoRecurse: Attribute = Attribute{bits: 1 << 48,}

Unstable (rustc_private)

const InaccessibleMemOnly: Attribute = Attribute{bits: 1 << 49,}

Unstable (rustc_private)

const InaccessibleMemOrArgMemOnly: Attribute = Attribute{bits: 1 << 50,}

Unstable (rustc_private)

fn empty() -> Attribute

Unstable (rustc_private)

Returns an empty set of flags.

fn all() -> Attribute

Unstable (rustc_private)

Returns the set containing all flags.

fn bits(&self) -> u64

Unstable (rustc_private)

Returns the raw value of the flags currently stored.

fn from_bits(bits: u64) -> Option<Attribute>

Unstable (rustc_private)

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: u64) -> Attribute

Unstable (rustc_private)

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Unstable (rustc_private)

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Unstable (rustc_private)

Returns true if all flags are currently set.

fn intersects(&self, other: Attribute) -> bool

Unstable (rustc_private)

Returns true if there are flags common to both self and other.

fn contains(&self, other: Attribute) -> bool

Unstable (rustc_private)

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: Attribute)

Unstable (rustc_private)

Inserts the specified flags in-place.

fn remove(&mut self, other: Attribute)

Unstable (rustc_private)

Removes the specified flags in-place.

fn toggle(&mut self, other: Attribute)

Unstable (rustc_private)

Toggles the specified flags in-place.