Crate rustc_llvm

Structs

RustStringInner 🔒

Underlying implementation of RustString.

Functions

LLVMRustStringWriteImpl⚠

Appends the contents of a byte slice to a RustString.

initialize_available_targets

Initialize targets enabled by the build script via cfg(llvm_component = "..."). N.B., this function can’t be moved to rustc_codegen_llvm because of the cfgs.

Foreign Types

RustString

Opaque type that allows C++ code to write bytes to a Rust-side buffer, in conjunction with RawRustStringOstream. Use this as &RustString (Rust) and RustStringRef (C++) in FFI signatures.