Crate rustc_llvm

Unstable (rustc_private)

Reexports

pub use self::AttributeSet::*;

pub use self::IntPredicate::*;

pub use self::RealPredicate::*;

pub use self::TypeKind::*;

pub use self::AtomicBinOp::*;

pub use self::AtomicOrdering::*;

pub use self::SynchronizationScope::*;

pub use self::FileType::*;

pub use self::MetadataType::*;

pub use self::AsmDialect::*;

pub use self::CodeGenOptLevel::*;

pub use self::CodeGenOptSize::*;

pub use self::RelocMode::*;

pub use self::CodeGenModel::*;

pub use self::DiagnosticKind::*;

pub use self::CallConv::*;

pub use self::Visibility::*;

pub use self::DiagnosticSeverity::*;

pub use self::Linkage::*;

pub use self::DLLStorageClassTypes::*;

Modules

archive_ro 	 [Unstable]
A wrapper around LLVM's archive (.a) code

debuginfo 	 [Unstable]

diagnostic 	 [Unstable]
LLVM diagnostic reports.

Structs

Attribute 	 [Unstable]

Attributes 	 [Unstable]

ObjectFile 	 [Unstable]

OperandBundleDef 	 [Unstable]

SectionIter 	 [Unstable]

TargetData 	 [Unstable]

Enums

ArchiveChild_opaque 	 [Unstable]

ArchiveIterator_opaque 	 [Unstable]

ArchiveKind 	 [Unstable]

Archive_opaque 	 [Unstable]

AsmDialect 	 [Unstable]

AtomicBinOp 	 [Unstable]

AtomicOrdering 	 [Unstable]

AttributeSet 	 [Unstable]

BasicBlock_opaque 	 [Unstable]

Builder_opaque 	 [Unstable]

CallConv 	 [Unstable]

CodeGenModel 	 [Unstable]

CodeGenOptLevel 	 [Unstable]

CodeGenOptSize 	 [Unstable]

Context_opaque 	 [Unstable]

DLLStorageClassTypes 	 [Unstable]

DebugLoc_opaque 	 [Unstable]

DiagnosticInfo_opaque 	 [Unstable]

DiagnosticKind 	 [Unstable]

DiagnosticSeverity 	 [Unstable]

ExecutionEngine_opaque 	 [Unstable]

FileType 	 [Unstable]

IntPredicate 	 [Unstable]

Linkage 	 [Unstable]

MemoryBuffer_opaque 	 [Unstable]

MetadataType 	 [Unstable]

Metadata_opaque 	 [Unstable]

Module_opaque 	 [Unstable]

ObjectFile_opaque 	 [Unstable]

OperandBundleDef_opaque 	 [Unstable]

PassManagerBuilder_opaque 	 [Unstable]

PassManager_opaque 	 [Unstable]

Pass_opaque 	 [Unstable]

RealPredicate 	 [Unstable]

RelocMode 	 [Unstable]

RustArchiveMember_opaque 	 [Unstable]

RustString_opaque 	 [Unstable]

SMDiagnostic_opaque 	 [Unstable]

SectionIterator_opaque 	 [Unstable]

SupportedPassKind 	 [Unstable]
Represents the different LLVM passes Rust supports

SynchronizationScope 	 [Unstable]

TargetData_opaque 	 [Unstable]

TargetMachine_opaque 	 [Unstable]

Twine_opaque 	 [Unstable]

TypeKind 	 [Unstable]

Type_opaque 	 [Unstable]

Use_opaque 	 [Unstable]

Value_opaque 	 [Unstable]

Visibility 	 [Unstable]

Constants

False 	 [Unstable]

True 	 [Unstable]

Functions

ConstFCmp 	 [Unstable]

ConstICmp 	 [Unstable]

LLVMABIAlignmentOfType 	 [Unstable]
Returns the minimum alignment of a type.

LLVMABISizeOfType 	 [Unstable]
Distance between successive elements in an array of T. Includes ABI padding.

LLVMAddAggressiveDCEPass 	 [Unstable]

LLVMAddAlias 	 [Unstable]

LLVMAddArgumentPromotionPass 	 [Unstable]

LLVMAddAttribute 	 [Unstable]

LLVMAddBasicAliasAnalysisPass 	 [Unstable]

LLVMAddCFGSimplificationPass 	 [Unstable]

LLVMAddCallSiteAttribute 	 [Unstable]

LLVMAddCase 	 [Unstable]

LLVMAddClause 	 [Unstable]

LLVMAddConstantMergePass 	 [Unstable]

LLVMAddConstantPropagationPass 	 [Unstable]

LLVMAddCorrelatedValuePropagationPass 	 [Unstable]

LLVMAddDeadArgEliminationPass 	 [Unstable]

LLVMAddDeadStoreEliminationPass 	 [Unstable]

LLVMAddDereferenceableAttr 	 [Unstable]

LLVMAddDereferenceableCallSiteAttr 	 [Unstable]

LLVMAddDestination 	 [Unstable]

LLVMAddEarlyCSEPass 	 [Unstable]

LLVMAddFunction 	 [Unstable]

LLVMAddFunctionAttrString 	 [Unstable]

LLVMAddFunctionAttrStringValue 	 [Unstable]

LLVMAddFunctionAttribute 	 [Unstable]

LLVMAddFunctionAttrsPass 	 [Unstable]

LLVMAddFunctionInliningPass 	 [Unstable]

LLVMAddGVNPass 	 [Unstable]

LLVMAddGlobal 	 [Unstable]

LLVMAddGlobalDCEPass 	 [Unstable]

LLVMAddGlobalInAddressSpace 	 [Unstable]

LLVMAddGlobalOptimizerPass 	 [Unstable]

LLVMAddIPSCCPPass 	 [Unstable]

LLVMAddIncoming 	 [Unstable]

LLVMAddIndVarSimplifyPass 	 [Unstable]

LLVMAddInstrAttribute 	 [Unstable]

LLVMAddInstructionCombiningPass 	 [Unstable]

LLVMAddJumpThreadingPass 	 [Unstable]

LLVMAddLICMPass 	 [Unstable]

LLVMAddLoopDeletionPass 	 [Unstable]

LLVMAddLoopIdiomPass 	 [Unstable]

LLVMAddLoopRotatePass 	 [Unstable]

LLVMAddLoopUnrollPass 	 [Unstable]

LLVMAddLoopUnswitchPass 	 [Unstable]

LLVMAddMemCpyOptPass 	 [Unstable]

LLVMAddNamedMetadataOperand 	 [Unstable]

LLVMAddPruneEHPass 	 [Unstable]

LLVMAddReassociatePass 	 [Unstable]

LLVMAddSCCPPass 	 [Unstable]

LLVMAddScalarReplAggregatesPass 	 [Unstable]

LLVMAddScalarReplAggregatesPassSSA 	 [Unstable]

LLVMAddSimplifyLibCallsPass 	 [Unstable]

LLVMAddStripDeadPrototypesPass 	 [Unstable]

LLVMAddTailCallEliminationPass 	 [Unstable]

LLVMAddTargetData 	 [Unstable]
Adds the target data to the given pass manager. The pass manager references the target data only weakly.

LLVMAddTypeBasedAliasAnalysisPass 	 [Unstable]

LLVMAddVerifierPass 	 [Unstable]
Adds a verification pass.

LLVMAlignOf 	 [Unstable]

LLVMAppendBasicBlockInContext 	 [Unstable]

LLVMBasicBlockAsValue 	 [Unstable]

LLVMBlockAddress 	 [Unstable]

LLVMBuildAShr 	 [Unstable]

LLVMBuildAdd 	 [Unstable]

LLVMBuildAggregateRet 	 [Unstable]

LLVMBuildAlloca 	 [Unstable]

LLVMBuildAnd 	 [Unstable]

LLVMBuildAtomicCmpXchg 	 [Unstable]

LLVMBuildAtomicFence 	 [Unstable]

LLVMBuildAtomicLoad 	 [Unstable]

LLVMBuildAtomicRMW 	 [Unstable]

LLVMBuildAtomicStore 	 [Unstable]

LLVMBuildBinOp 	 [Unstable]

LLVMBuildBitCast 	 [Unstable]

LLVMBuildBr 	 [Unstable]

LLVMBuildCast 	 [Unstable]

LLVMBuildCondBr 	 [Unstable]

LLVMBuildExactSDiv 	 [Unstable]

LLVMBuildExecutionEngine 	 [Unstable]

LLVMBuildExtractElement 	 [Unstable]

LLVMBuildExtractValue 	 [Unstable]

LLVMBuildFAdd 	 [Unstable]

LLVMBuildFCmp 	 [Unstable]

LLVMBuildFDiv 	 [Unstable]

LLVMBuildFMul 	 [Unstable]

LLVMBuildFNeg 	 [Unstable]

LLVMBuildFPCast 	 [Unstable]

LLVMBuildFPExt 	 [Unstable]

LLVMBuildFPToSI 	 [Unstable]

LLVMBuildFPToUI 	 [Unstable]

LLVMBuildFPTrunc 	 [Unstable]

LLVMBuildFRem 	 [Unstable]

LLVMBuildFSub 	 [Unstable]

LLVMBuildFree 	 [Unstable]

LLVMBuildGEP 	 [Unstable]

LLVMBuildGlobalString 	 [Unstable]

LLVMBuildGlobalStringPtr 	 [Unstable]

LLVMBuildICmp 	 [Unstable]

LLVMBuildInBoundsGEP 	 [Unstable]

LLVMBuildIndirectBr 	 [Unstable]

LLVMBuildInsertElement 	 [Unstable]

LLVMBuildInsertValue 	 [Unstable]

LLVMBuildIntCast 	 [Unstable]

LLVMBuildIntToPtr 	 [Unstable]

LLVMBuildIsNotNull 	 [Unstable]

LLVMBuildIsNull 	 [Unstable]

LLVMBuildLShr 	 [Unstable]

LLVMBuildLoad 	 [Unstable]

LLVMBuildMul 	 [Unstable]

LLVMBuildNSWAdd 	 [Unstable]

LLVMBuildNSWMul 	 [Unstable]

LLVMBuildNSWNeg 	 [Unstable]

LLVMBuildNSWSub 	 [Unstable]

LLVMBuildNUWAdd 	 [Unstable]

LLVMBuildNUWMul 	 [Unstable]

LLVMBuildNUWNeg 	 [Unstable]

LLVMBuildNUWSub 	 [Unstable]

LLVMBuildNeg 	 [Unstable]

LLVMBuildNot 	 [Unstable]

LLVMBuildOr 	 [Unstable]

LLVMBuildPhi 	 [Unstable]

LLVMBuildPointerCast 	 [Unstable]

LLVMBuildPtrDiff 	 [Unstable]

LLVMBuildPtrToInt 	 [Unstable]

LLVMBuildResume 	 [Unstable]

LLVMBuildRet 	 [Unstable]

LLVMBuildRetVoid 	 [Unstable]

LLVMBuildSDiv 	 [Unstable]

LLVMBuildSExt 	 [Unstable]

LLVMBuildSExtOrBitCast 	 [Unstable]

LLVMBuildSIToFP 	 [Unstable]

LLVMBuildSRem 	 [Unstable]

LLVMBuildSelect 	 [Unstable]

LLVMBuildShl 	 [Unstable]

LLVMBuildShuffleVector 	 [Unstable]

LLVMBuildStore 	 [Unstable]

LLVMBuildStructGEP 	 [Unstable]

LLVMBuildSub 	 [Unstable]

LLVMBuildSwitch 	 [Unstable]

LLVMBuildTrunc 	 [Unstable]

LLVMBuildTruncOrBitCast 	 [Unstable]

LLVMBuildUDiv 	 [Unstable]

LLVMBuildUIToFP 	 [Unstable]

LLVMBuildURem 	 [Unstable]

LLVMBuildUnreachable 	 [Unstable]

LLVMBuildVAArg 	 [Unstable]

LLVMBuildXor 	 [Unstable]

LLVMBuildZExt 	 [Unstable]

LLVMBuildZExtOrBitCast 	 [Unstable]

LLVMCallFrameAlignmentOfType 	 [Unstable]
Returns the minimum alignment of a type when part of a call frame.

LLVMClearInsertionPosition 	 [Unstable]

LLVMCloneModule 	 [Unstable]

LLVMConstAShr 	 [Unstable]

LLVMConstAdd 	 [Unstable]

LLVMConstAllOnes 	 [Unstable]

LLVMConstAnd 	 [Unstable]

LLVMConstArray 	 [Unstable]

LLVMConstBitCast 	 [Unstable]

LLVMConstExactSDiv 	 [Unstable]

LLVMConstExtractElement 	 [Unstable]

LLVMConstExtractValue 	 [Unstable]

LLVMConstFAdd 	 [Unstable]

LLVMConstFCmp 	 [Unstable]

LLVMConstFDiv 	 [Unstable]

LLVMConstFMul 	 [Unstable]

LLVMConstFNeg 	 [Unstable]

LLVMConstFPCast 	 [Unstable]

LLVMConstFPExt 	 [Unstable]

LLVMConstFPToSI 	 [Unstable]

LLVMConstFPToUI 	 [Unstable]

LLVMConstFPTrunc 	 [Unstable]

LLVMConstFRem 	 [Unstable]

LLVMConstFSub 	 [Unstable]

LLVMConstGEP 	 [Unstable]

LLVMConstICmp 	 [Unstable]

LLVMConstInBoundsGEP 	 [Unstable]

LLVMConstInlineAsm 	 [Unstable]

LLVMConstInsertElement 	 [Unstable]

LLVMConstInsertValue 	 [Unstable]

LLVMConstInt 	 [Unstable]

LLVMConstIntCast 	 [Unstable]

LLVMConstIntGetSExtValue 	 [Unstable]

LLVMConstIntGetZExtValue 	 [Unstable]

LLVMConstIntOfString 	 [Unstable]

LLVMConstIntOfStringAndSize 	 [Unstable]

LLVMConstIntToPtr 	 [Unstable]

LLVMConstLShr 	 [Unstable]

LLVMConstMul 	 [Unstable]

LLVMConstNSWAdd 	 [Unstable]

LLVMConstNSWMul 	 [Unstable]

LLVMConstNSWNeg 	 [Unstable]

LLVMConstNSWSub 	 [Unstable]

LLVMConstNUWAdd 	 [Unstable]

LLVMConstNUWMul 	 [Unstable]

LLVMConstNUWNeg 	 [Unstable]

LLVMConstNUWSub 	 [Unstable]

LLVMConstNamedStruct 	 [Unstable]

LLVMConstNeg 	 [Unstable]

LLVMConstNot 	 [Unstable]

LLVMConstNull 	 [Unstable]

LLVMConstOr 	 [Unstable]

LLVMConstPointerCast 	 [Unstable]

LLVMConstPointerNull 	 [Unstable]

LLVMConstPtrToInt 	 [Unstable]

LLVMConstReal 	 [Unstable]

LLVMConstRealOfString 	 [Unstable]

LLVMConstRealOfStringAndSize 	 [Unstable]

LLVMConstSDiv 	 [Unstable]

LLVMConstSExt 	 [Unstable]

LLVMConstSExtOrBitCast 	 [Unstable]

LLVMConstSIToFP 	 [Unstable]

LLVMConstSRem 	 [Unstable]

LLVMConstSelect 	 [Unstable]

LLVMConstShl 	 [Unstable]

LLVMConstShuffleVector 	 [Unstable]

LLVMConstStringInContext 	 [Unstable]

LLVMConstStructInContext 	 [Unstable]

LLVMConstSub 	 [Unstable]

LLVMConstTrunc 	 [Unstable]

LLVMConstTruncOrBitCast 	 [Unstable]

LLVMConstUDiv 	 [Unstable]

LLVMConstUIToFP 	 [Unstable]

LLVMConstURem 	 [Unstable]

LLVMConstVector 	 [Unstable]

LLVMConstXor 	 [Unstable]

LLVMConstZExt 	 [Unstable]

LLVMConstZExtOrBitCast 	 [Unstable]

LLVMContextCreate 	 [Unstable]

LLVMContextDispose 	 [Unstable]

LLVMContextSetDiagnosticHandler 	 [Unstable]

LLVMCountBasicBlocks 	 [Unstable]

LLVMCountIncoming 	 [Unstable]

LLVMCountParamTypes 	 [Unstable]

LLVMCountParams 	 [Unstable]

LLVMCountStructElementTypes 	 [Unstable]

LLVMCreateBuilderInContext 	 [Unstable]

LLVMCreateFunctionPassManagerForModule 	 [Unstable]
Creates a function-by-function pass manager

LLVMCreateMemoryBufferWithMemoryRange 	 [Unstable]
Borrows the contents of the memory buffer (doesn't copy it)

LLVMCreateMemoryBufferWithMemoryRangeCopy 	 [Unstable]

LLVMCreateObjectFile 	 [Unstable]
Opens an object file.

LLVMCreatePassManager 	 [Unstable]
Creates a pass manager.

LLVMCreateTargetData 	 [Unstable]
Creates target data from a target layout string.

LLVMDIBuilderCreate 	 [Unstable]

LLVMDIBuilderCreateArrayType 	 [Unstable]

LLVMDIBuilderCreateBasicType 	 [Unstable]

LLVMDIBuilderCreateCompileUnit 	 [Unstable]

LLVMDIBuilderCreateDebugLocation 	 [Unstable]

LLVMDIBuilderCreateEnumerationType 	 [Unstable]

LLVMDIBuilderCreateEnumerator 	 [Unstable]

LLVMDIBuilderCreateFile 	 [Unstable]

LLVMDIBuilderCreateFunction 	 [Unstable]

LLVMDIBuilderCreateLexicalBlock 	 [Unstable]

LLVMDIBuilderCreateMemberType 	 [Unstable]

LLVMDIBuilderCreateNameSpace 	 [Unstable]

LLVMDIBuilderCreateOpDeref 	 [Unstable]

LLVMDIBuilderCreateOpPlus 	 [Unstable]

LLVMDIBuilderCreatePointerType 	 [Unstable]

LLVMDIBuilderCreateStaticVariable 	 [Unstable]

LLVMDIBuilderCreateStructType 	 [Unstable]

LLVMDIBuilderCreateSubroutineType 	 [Unstable]

LLVMDIBuilderCreateTemplateTypeParameter 	 [Unstable]

LLVMDIBuilderCreateUnionType 	 [Unstable]

LLVMDIBuilderCreateVariable 	 [Unstable]

LLVMDIBuilderCreateVectorType 	 [Unstable]

LLVMDIBuilderDispose 	 [Unstable]

LLVMDIBuilderFinalize 	 [Unstable]

LLVMDIBuilderGetOrCreateArray 	 [Unstable]

LLVMDIBuilderGetOrCreateSubrange 	 [Unstable]

LLVMDIBuilderInsertDeclareAtEnd 	 [Unstable]

LLVMDIBuilderInsertDeclareBefore 	 [Unstable]

LLVMDICompositeTypeSetTypeArray 	 [Unstable]

LLVMDeleteBasicBlock 	 [Unstable]

LLVMDeleteFunction 	 [Unstable]

LLVMDeleteGlobal 	 [Unstable]

LLVMDisposeBuilder 	 [Unstable]

LLVMDisposeExecutionEngine 	 [Unstable]

LLVMDisposeMemoryBuffer 	 [Unstable]
Destroys a memory buffer.

LLVMDisposeModule 	 [Unstable]

LLVMDisposeObjectFile 	 [Unstable]
Closes an object file.

LLVMDisposePassManager 	 [Unstable]
Disposes a pass manager.

LLVMDisposeSectionIterator 	 [Unstable]
Destroys a section iterator.

LLVMDisposeTargetData 	 [Unstable]
Disposes target data.

LLVMDoubleTypeInContext 	 [Unstable]

LLVMDumpModule 	 [Unstable]
See Module::dump.

LLVMDumpValue 	 [Unstable]

LLVMExecutionEngineAddModule 	 [Unstable]

LLVMExecutionEngineFinalizeObject 	 [Unstable]

LLVMExecutionEngineRemoveModule 	 [Unstable]

LLVMFP128TypeInContext 	 [Unstable]

LLVMFinalizeFunctionPassManager 	 [Unstable]
Finalizes all the function passes scheduled in the manager

LLVMFloatTypeInContext 	 [Unstable]

LLVMFunctionType 	 [Unstable]

LLVMGetAlignment 	 [Unstable]

LLVMGetArrayLength 	 [Unstable]

LLVMGetAttribute 	 [Unstable]

LLVMGetBasicBlockParent 	 [Unstable]

LLVMGetBasicBlocks 	 [Unstable]

LLVMGetCurrentDebugLocation 	 [Unstable]

LLVMGetDataLayout 	 [Unstable]
Data layout. See Module::getDataLayout.

LLVMGetDiagInfoKind 	 [Unstable]

LLVMGetDiagInfoSeverity 	 [Unstable]

LLVMGetElementType 	 [Unstable]

LLVMGetEntryBasicBlock 	 [Unstable]

LLVMGetFirstBasicBlock 	 [Unstable]

LLVMGetFirstFunction 	 [Unstable]

LLVMGetFirstGlobal 	 [Unstable]

LLVMGetFirstInstruction 	 [Unstable]

LLVMGetFirstParam 	 [Unstable]

LLVMGetFirstUse 	 [Unstable]

LLVMGetFunctionAttr 	 [Unstable]

LLVMGetFunctionCallConv 	 [Unstable]

LLVMGetGC 	 [Unstable]

LLVMGetGlobalParent 	 [Unstable]

LLVMGetIncomingBlock 	 [Unstable]

LLVMGetIncomingValue 	 [Unstable]

LLVMGetInitializer 	 [Unstable]

LLVMGetInsertBlock 	 [Unstable]

LLVMGetInstructionCallConv 	 [Unstable]

LLVMGetInstructionParent 	 [Unstable]

LLVMGetIntTypeWidth 	 [Unstable]

LLVMGetIntrinsicID 	 [Unstable]

LLVMGetLastBasicBlock 	 [Unstable]

LLVMGetLastFunction 	 [Unstable]

LLVMGetLastGlobal 	 [Unstable]

LLVMGetLastInstruction 	 [Unstable]

LLVMGetLastParam 	 [Unstable]

LLVMGetLinkage 	 [Unstable]

LLVMGetMDKindIDInContext 	 [Unstable]

LLVMGetMetadata 	 [Unstable]

LLVMGetModuleContext 	 [Unstable]

LLVMGetNamedFunction 	 [Unstable]

LLVMGetNamedGlobal 	 [Unstable]

LLVMGetNamedValue 	 [Unstable]

LLVMGetNextBasicBlock 	 [Unstable]

LLVMGetNextFunction 	 [Unstable]

LLVMGetNextGlobal 	 [Unstable]

LLVMGetNextInstruction 	 [Unstable]

LLVMGetNextParam 	 [Unstable]

LLVMGetNextUse 	 [Unstable]

LLVMGetNumOperands 	 [Unstable]

LLVMGetOperand 	 [Unstable]

LLVMGetOrInsertFunction 	 [Unstable]

LLVMGetOrInsertGlobal 	 [Unstable]

LLVMGetParam 	 [Unstable]

LLVMGetParamParent 	 [Unstable]

LLVMGetParamTypes 	 [Unstable]

LLVMGetParams 	 [Unstable]

LLVMGetPointerAddressSpace 	 [Unstable]

LLVMGetPointerToGlobal 	 [Unstable]

LLVMGetPreviousBasicBlock 	 [Unstable]

LLVMGetPreviousFunction 	 [Unstable]

LLVMGetPreviousGlobal 	 [Unstable]

LLVMGetPreviousInstruction 	 [Unstable]

LLVMGetPreviousParam 	 [Unstable]

LLVMGetReturnType 	 [Unstable]

LLVMGetSection 	 [Unstable]

LLVMGetSectionContents 	 [Unstable]
Returns the current section contents as a string buffer.

LLVMGetSectionSize 	 [Unstable]
Returns the current section size.

LLVMGetSections 	 [Unstable]
Enumerates the sections in an object file.

LLVMGetStructElementTypes 	 [Unstable]

LLVMGetTarget 	 [Unstable]
Target triple. See Module::getTargetTriple.

LLVMGetTypeContext 	 [Unstable]
See llvm::LLVMType::getContext.

LLVMGetTypeKind 	 [Unstable]
See llvm::LLVMTypeKind::getTypeID.

LLVMGetUndef 	 [Unstable]

LLVMGetUsedValue 	 [Unstable]

LLVMGetUser 	 [Unstable]

LLVMGetValueName 	 [Unstable]

LLVMGetVectorSize 	 [Unstable]

LLVMGetVisibility 	 [Unstable]

LLVMGetVolatile 	 [Unstable]

LLVMHasMetadata 	 [Unstable]

LLVMInitializeFunctionPassManager 	 [Unstable]
Initializes all the function passes scheduled in the manager

LLVMInitializePasses 	 [Unstable]

LLVMInlineAsm 	 [Unstable]
Prepares inline assembly.

LLVMInsertBasicBlockInContext 	 [Unstable]

LLVMInsertIntoBuilder 	 [Unstable]

LLVMInsertIntoBuilderWithName 	 [Unstable]

LLVMInstructionEraseFromParent 	 [Unstable]

LLVMInt16TypeInContext 	 [Unstable]

LLVMInt1TypeInContext 	 [Unstable]

LLVMInt32TypeInContext 	 [Unstable]

LLVMInt64TypeInContext 	 [Unstable]

LLVMInt8TypeInContext 	 [Unstable]

LLVMIntTypeInContext 	 [Unstable]

LLVMIsAAllocaInst 	 [Unstable]

LLVMIsAArgument 	 [Unstable]

LLVMIsAConstantInt 	 [Unstable]

LLVMIsAGlobalVariable 	 [Unstable]

LLVMIsAStoreInst 	 [Unstable]

LLVMIsATerminatorInst 	 [Unstable]

LLVMIsConstant 	 [Unstable]

LLVMIsDeclaration 	 [Unstable]

LLVMIsFunctionVarArg 	 [Unstable]

LLVMIsGlobalConstant 	 [Unstable]

LLVMIsMultithreaded 	 [Unstable]

LLVMIsNull 	 [Unstable]

LLVMIsPackedStruct 	 [Unstable]

LLVMIsSectionIteratorAtEnd 	 [Unstable]
Returns true if the section iterator is at the end of the section list:

LLVMIsTailCall 	 [Unstable]

LLVMIsThreadLocal 	 [Unstable]

LLVMIsUndef 	 [Unstable]

LLVMLabelTypeInContext 	 [Unstable]

LLVMMDNodeInContext 	 [Unstable]

LLVMMDStringInContext 	 [Unstable]

LLVMMetadataTypeInContext 	 [Unstable]

LLVMModuleCreateWithNameInContext 	 [Unstable]

LLVMMoveBasicBlockAfter 	 [Unstable]

LLVMMoveBasicBlockBefore 	 [Unstable]

LLVMMoveToNextSection 	 [Unstable]
Moves the section iterator to point to the next section.

LLVMOffsetOfElement 	 [Unstable]
Computes the byte offset of the indexed struct element for a target.

LLVMPPCFP128TypeInContext 	 [Unstable]

LLVMPassManagerBuilderCreate 	 [Unstable]

LLVMPassManagerBuilderDispose 	 [Unstable]

LLVMPassManagerBuilderPopulateFunctionPassManager 	 [Unstable]

LLVMPassManagerBuilderPopulateLTOPassManager 	 [Unstable]

LLVMPassManagerBuilderPopulateModulePassManager 	 [Unstable]

LLVMPassManagerBuilderSetDisableSimplifyLibCalls 	 [Unstable]

LLVMPassManagerBuilderSetDisableUnitAtATime 	 [Unstable]

LLVMPassManagerBuilderSetDisableUnrollLoops 	 [Unstable]

LLVMPassManagerBuilderSetOptLevel 	 [Unstable]

LLVMPassManagerBuilderSetSizeLevel 	 [Unstable]

LLVMPassManagerBuilderUseInlinerWithThreshold 	 [Unstable]

LLVMPointerType 	 [Unstable]

LLVMPositionBuilder 	 [Unstable]

LLVMPositionBuilderAtEnd 	 [Unstable]

LLVMPositionBuilderBefore 	 [Unstable]

LLVMPreferredAlignmentOfType 	 [Unstable]
Returns the preferred alignment of a type.

LLVMRemoveAttribute 	 [Unstable]

LLVMRemoveFunctionAttr 	 [Unstable]

LLVMRemoveFunctionAttrString 	 [Unstable]

LLVMRemoveFunctionAttributes 	 [Unstable]

LLVMRemoveInstrAttribute 	 [Unstable]

LLVMReplaceAllUsesWith 	 [Unstable]

LLVMRunFunctionPassManager 	 [Unstable]
Runs the function passes on the provided function.

LLVMRunPassManager 	 [Unstable]
Runs a pass manager on a module.

LLVMRustAddAlwaysInlinePass 	 [Unstable]

LLVMRustAddAnalysisPasses 	 [Unstable]

LLVMRustAddBuilderLibraryInfo 	 [Unstable]

LLVMRustAddHandler 	 [Unstable]

LLVMRustAddLibraryInfo 	 [Unstable]

LLVMRustAddModuleFlag 	 [Unstable]

LLVMRustAddPass 	 [Unstable]

LLVMRustArchiveChildData 	 [Unstable]

LLVMRustArchiveChildFree 	 [Unstable]

LLVMRustArchiveChildName 	 [Unstable]

LLVMRustArchiveIteratorFree 	 [Unstable]

LLVMRustArchiveIteratorNew 	 [Unstable]

LLVMRustArchiveIteratorNext 	 [Unstable]

LLVMRustArchiveMemberFree 	 [Unstable]

LLVMRustArchiveMemberNew 	 [Unstable]

LLVMRustArrayType 	 [Unstable]

LLVMRustBuildCall 	 [Unstable]

LLVMRustBuildCatchPad 	 [Unstable]

LLVMRustBuildCatchRet 	 [Unstable]

LLVMRustBuildCatchSwitch 	 [Unstable]

LLVMRustBuildCleanupPad 	 [Unstable]

LLVMRustBuildCleanupRet 	 [Unstable]

LLVMRustBuildInvoke 	 [Unstable]

LLVMRustBuildLandingPad 	 [Unstable]

LLVMRustBuildOperandBundleDef 	 [Unstable]

LLVMRustConfigurePassManagerBuilder 	 [Unstable]

LLVMRustCreateMemoryBufferWithContentsOfFile 	 [Unstable]
Reads the given file and returns it as a memory buffer. Use LLVMDisposeMemoryBuffer() to get rid of it.

LLVMRustCreateTargetMachine 	 [Unstable]

LLVMRustDebugMetadataVersion 	 [Unstable]

LLVMRustDestroyArchive 	 [Unstable]

LLVMRustDisposeTargetMachine 	 [Unstable]

LLVMRustFindAndCreatePass 	 [Unstable]

LLVMRustFreeOperandBundleDef 	 [Unstable]

LLVMRustGetLastError 	 [Unstable]
Returns a string describing the last error caused by an LLVMRust* call.

LLVMRustGetModuleDataLayout 	 [Unstable]

LLVMRustGetSectionName 	 [Unstable]

LLVMRustHasFeature 	 [Unstable]

LLVMRustLinkInExternalBitcode 	 [Unstable]

LLVMRustLoadDynamicLibrary 	 [Unstable]

LLVMRustMarkAllFunctionsNounwind 	 [Unstable]

LLVMRustOpenArchive 	 [Unstable]

LLVMRustPassKind 	 [Unstable]

LLVMRustPositionBuilderAtStart 	 [Unstable]

LLVMRustPrintModule 	 [Unstable]

LLVMRustPrintPassTimings 	 [Unstable]
Print the pass timings since static dtors aren't picking them up.

LLVMRustPrintPasses 	 [Unstable]

LLVMRustRunFunctionPassManager 	 [Unstable]

LLVMRustRunRestrictionPass 	 [Unstable]

LLVMRustSetComdat 	 [Unstable]

LLVMRustSetDLLStorageClass 	 [Unstable]

LLVMRustSetDataLayoutFromTargetMachine 	 [Unstable]

LLVMRustSetHasUnsafeAlgebra 	 [Unstable]

LLVMRustSetLLVMOptions 	 [Unstable]

LLVMRustSetNormalizedTarget 	 [Unstable]

LLVMRustSetPersonalityFn 	 [Unstable]

LLVMRustUnsetComdat 	 [Unstable]

LLVMRustWriteArchive 	 [Unstable]

LLVMRustWriteOutputFile 	 [Unstable]

LLVMSetAlignment 	 [Unstable]

LLVMSetCleanup 	 [Unstable]

LLVMSetCurrentDebugLocation 	 [Unstable]

LLVMSetDataLayout 	 [Unstable]

LLVMSetDebug 	 [Unstable]
Enables LLVM debug output.

LLVMSetFunctionCallConv 	 [Unstable]

LLVMSetGC 	 [Unstable]

LLVMSetGlobalConstant 	 [Unstable]

LLVMSetInitializer 	 [Unstable]

LLVMSetInlineAsmDiagnosticHandler 	 [Unstable]

LLVMSetInstDebugLocation 	 [Unstable]

LLVMSetInstrParamAlignment 	 [Unstable]

LLVMSetInstructionCallConv 	 [Unstable]

LLVMSetLinkage 	 [Unstable]

LLVMSetMetadata 	 [Unstable]

LLVMSetModuleInlineAsm 	 [Unstable]
See Module::setModuleInlineAsm.

LLVMSetOperand 	 [Unstable]

LLVMSetParamAlignment 	 [Unstable]

LLVMSetSection 	 [Unstable]

LLVMSetTailCall 	 [Unstable]

LLVMSetTarget 	 [Unstable]

LLVMSetThreadLocal 	 [Unstable]

LLVMSetUnnamedAddr 	 [Unstable]

LLVMSetValueName 	 [Unstable]

LLVMSetVisibility 	 [Unstable]

LLVMSetVolatile 	 [Unstable]

LLVMSizeOf 	 [Unstable]

LLVMSizeOfTypeInBits 	 [Unstable]
Number of bytes clobbered when doing a Store to *T.

LLVMStartMultithreaded 	 [Unstable]

LLVMStoreSizeOfType 	 [Unstable]
Number of bytes clobbered when doing a Store to *T.

LLVMStructCreateNamed 	 [Unstable]

LLVMStructSetBody 	 [Unstable]

LLVMStructTypeInContext 	 [Unstable]

LLVMTypeOf 	 [Unstable]

LLVMUnpackInlineAsmDiagnostic 	 [Unstable]

LLVMUnpackOptimizationDiagnostic 	 [Unstable]

LLVMValueAsBasicBlock 	 [Unstable]

LLVMValueIsBasicBlock 	 [Unstable]

LLVMVectorType 	 [Unstable]

LLVMVersionMajor 	 [Unstable]

LLVMVersionMinor 	 [Unstable]

LLVMVoidTypeInContext 	 [Unstable]

LLVMWriteBitcodeToFile 	 [Unstable]
Writes a module to the specified path. Returns 0 on success.

LLVMWriteDebugLocToString 	 [Unstable]

LLVMWriteDiagnosticInfoToString 	 [Unstable]

LLVMWriteSMDiagnosticToString 	 [Unstable]

LLVMWriteTwineToString 	 [Unstable]

LLVMWriteTypeToString 	 [Unstable]

LLVMWriteValueToString 	 [Unstable]

LLVMX86FP80TypeInContext 	 [Unstable]

RemoveFunctionAttributes 	 [Unstable]

SetDLLStorageClass 	 [Unstable]

SetFunctionAttribute 	 [Unstable]

SetFunctionCallConv 	 [Unstable]

SetInstructionCallConv 	 [Unstable]

SetLinkage 	 [Unstable]

SetUniqueComdat 	 [Unstable]

SetUnnamedAddr 	 [Unstable]

UnsetComdat 	 [Unstable]

build_string 	 [Unstable]

debug_loc_to_string 	 [Unstable]

get_param 	 [Unstable]
Safe wrapper around LLVMGetParam, because segfaults are no fun.

get_params 	 [Unstable]

initialize_available_targets 	 [Unstable]

last_error 	 [Unstable]

mk_section_iter 	 [Unstable]

mk_target_data 	 [Unstable]

rust_llvm_string_write_impl 	 [Unstable]
Appending to a Rust string -- used by raw_rust_string_ostream.

set_thread_local 	 [Unstable]

twine_to_string 	 [Unstable]

Type Definitions

ArchiveChildRef 	 [Unstable]

ArchiveIteratorRef 	 [Unstable]

ArchiveRef 	 [Unstable]

BasicBlockRef 	 [Unstable]

Bool 	 [Unstable]

BuilderRef 	 [Unstable]

ContextRef 	 [Unstable]

DebugLocRef 	 [Unstable]

DiagnosticHandler 	 [Unstable]

DiagnosticInfoRef 	 [Unstable]

ExecutionEngineRef 	 [Unstable]

InlineAsmDiagHandler 	 [Unstable]

MemoryBufferRef 	 [Unstable]

MetadataRef 	 [Unstable]

ModuleRef 	 [Unstable]

ObjectFileRef 	 [Unstable]

Opcode 	 [Unstable]

OperandBundleDefRef 	 [Unstable]

PassManagerBuilderRef 	 [Unstable]

PassManagerRef 	 [Unstable]

PassRef 	 [Unstable]

RustArchiveMemberRef 	 [Unstable]

RustStringRef 	 [Unstable]

SMDiagnosticRef 	 [Unstable]

SectionIteratorRef 	 [Unstable]

TargetDataRef 	 [Unstable]

TargetMachineRef 	 [Unstable]

TwineRef 	 [Unstable]

TypeRef 	 [Unstable]

UseRef 	 [Unstable]

ValueRef 	 [Unstable]