Struct rustc_llvm::Attribute

pub struct Attribute {
    // some fields omitted
}

Unstable (rustc_private)

Methods

impl Attribute

const ZExt: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const SExt: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const NoReturn: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const InReg: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const StructRet: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const NoUnwind: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const NoAlias: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const ByVal: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const Nest: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const ReadNone: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const ReadOnly: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const NoInline: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const AlwaysInline: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const OptimizeForSize: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const StackProtect: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const StackProtectReq: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const Alignment: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const NoCapture: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const NoRedZone: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const NoImplicitFloat: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const Naked: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const InlineHint: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const Stack: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const ReturnsTwice: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const UWTable: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const NonLazyBind: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

const OptimizeNone: Attribute = $ BitFlags { bits : $ value }

Unstable (rustc_private)

fn empty() -> Attribute

Unstable (rustc_private)

Returns an empty set of flags.

fn all() -> Attribute

Unstable (rustc_private)

Returns the set containing all flags.

fn bits(&self) -> u64

Unstable (rustc_private)

Returns the raw value of the flags currently stored.

fn from_bits(bits: u64) -> Option<Attribute>

Unstable (rustc_private)

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: u64) -> Attribute

Unstable (rustc_private)

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Unstable (rustc_private)

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Unstable (rustc_private)

Returns true if all flags are currently set.

fn intersects(&self, other: Attribute) -> bool

Unstable (rustc_private)

Returns true if there are flags common to both self and other.

fn contains(&self, other: Attribute) -> bool

Unstable (rustc_private)

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: Attribute)

Unstable (rustc_private)

Inserts the specified flags in-place.

fn remove(&mut self, other: Attribute)

Unstable (rustc_private)

Removes the specified flags in-place.

fn toggle(&mut self, other: Attribute)

Unstable (rustc_private)

Toggles the specified flags in-place.