Struct rustc_llvm::AttributeUnstable

pub struct Attribute {
    // some fields omitted
}

Methods

impl Attribute

fn empty() -> Attribute

Returns an empty set of flags.

fn all() -> Attribute

Returns the set containing all flags.

fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

fn from_bits(bits: u32) -> Option<Attribute>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: u32) -> Attribute

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently set.

fn intersects(&self, other: Attribute) -> bool

Returns true if there are flags common to both self and other.

fn contains(&self, other: Attribute) -> bool

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: Attribute)

Inserts the specified flags in-place.

fn remove(&mut self, other: Attribute)

Removes the specified flags in-place.

fn toggle(&mut self, other: Attribute)

Toggles the specified flags in-place.